refactor(nginx-module): split filter internals and metrics state
