+++ components/nginx-module/src/ngx_http_markdown_filter_module.c
@@ -3,9 +3,9 @@
  *
  * This module provides HTML to Markdown conversion for AI agents via
  * HTTP content negotiation (Accept: text/markdown).
- *
- * Phase 1: Basic module structure and configuration
- * Full implementation will be completed in subsequent phases.
+ * The main translation unit now mostly owns shared globals and module wiring,
+ * while configuration/bootstrap, request-path orchestration, worker lifecycle,
+ * and metrics endpoint helpers live in dedicated implementation includes.
  */
 
 #include "ngx_http_markdown_filter_module.h"
@@ -26,91 +26,34 @@
 /* Global converter instance for this worker process */
 static struct MarkdownConverterHandle *ngx_http_markdown_converter = NULL;
 
-/* Global metrics instance for this worker process */
-static ngx_http_markdown_metrics_t ngx_http_markdown_metrics;
+/* Global pointer to shared metrics state, resolved during worker init. */
+static ngx_http_markdown_metrics_t *ngx_http_markdown_metrics = NULL;
+static ngx_shm_zone_t *ngx_http_markdown_metrics_shm_zone = NULL;
+static ngx_str_t ngx_http_markdown_metrics_shm_name = ngx_string("nginx_markdown_metrics");
 static u_char ngx_http_markdown_empty_string[] = "";
+
+#define NGX_HTTP_MARKDOWN_METRIC_ADD(field, value)                                  \
+    do {                                                                            \
+        if (ngx_http_markdown_metrics != NULL) {                                    \
+            ngx_atomic_fetch_add(&ngx_http_markdown_metrics->field,                 \
+                (ngx_atomic_uint_t) (value));                                       \
+        }                                                                           \
+    } while (0)
+
+#define NGX_HTTP_MARKDOWN_METRIC_INC(field)                                         \
+    NGX_HTTP_MARKDOWN_METRIC_ADD(field, 1)
 
 /* Forward declarations */
 static ngx_int_t ngx_http_markdown_filter_init(ngx_conf_t *cf);
 static ngx_int_t ngx_http_markdown_init_worker(ngx_cycle_t *cycle);
 static void ngx_http_markdown_exit_worker(ngx_cycle_t *cycle);
-static void *ngx_http_markdown_create_conf(ngx_conf_t *cf);
-static char *ngx_http_markdown_merge_conf(ngx_conf_t *cf, void *parent, void *child);
 
-/* Configuration directive handlers */
-static char *ngx_http_markdown_filter(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_on_error(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_flavor(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_auth_policy(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_auth_cookies(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_conditional_requests(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_log_verbosity(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_stream_types(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-static char *ngx_http_markdown_metrics_directive(ngx_conf_t *cf, ngx_command_t *cmd, void *conf);
-
-/* Filter function declarations */
+/* Filter entrypoint declarations */
 static ngx_int_t ngx_http_markdown_header_filter(ngx_http_request_t *r);
 static ngx_int_t ngx_http_markdown_body_filter(ngx_http_request_t *r, ngx_chain_t *in);
-static ngx_int_t ngx_http_markdown_body_filter_buffer_input(
-    ngx_http_request_t *r, ngx_chain_t *in, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf);
-static ngx_int_t ngx_http_markdown_body_filter_decompress_if_needed(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf);
-static ngx_int_t ngx_http_markdown_body_filter_convert_and_output(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf);
-static ngx_int_t ngx_http_markdown_fail_open_buffered_response(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    const char *debug_message);
-static ngx_int_t ngx_http_markdown_reject_or_fail_open_buffered_response(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf, const char *debug_message);
-static ngx_int_t ngx_http_markdown_prepare_compressed_chain(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf, ngx_chain_t **compressed_chain);
-static ngx_int_t ngx_http_markdown_apply_decompressed_payload(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf, ngx_chain_t *decompressed_chain);
-static void ngx_http_markdown_record_decompression_success(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx);
-static ngx_int_t ngx_http_markdown_resolve_conditional_result(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf, struct MarkdownResult *result,
-    ngx_msec_t *elapsed_ms, ngx_flag_t *has_result);
-static ngx_int_t ngx_http_markdown_execute_conversion(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf, struct MarkdownResult *result,
-    ngx_msec_t *elapsed_ms);
-static ngx_int_t ngx_http_markdown_send_conversion_output(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf, struct MarkdownResult *result,
-    ngx_msec_t elapsed_ms);
 
 /* Metrics endpoint handler */
 static ngx_int_t ngx_http_markdown_metrics_handler(ngx_http_request_t *r);
-
-/* Configuration logging helpers */
-static ngx_uint_t ngx_http_markdown_log_verbosity_to_ngx_level(ngx_uint_t verbosity);
-static const ngx_str_t *ngx_http_markdown_on_error_name(ngx_uint_t value);
-static const ngx_str_t *ngx_http_markdown_flavor_name(ngx_uint_t value);
-static const ngx_str_t *ngx_http_markdown_auth_policy_name(ngx_uint_t value);
-static const ngx_str_t *ngx_http_markdown_conditional_requests_name(ngx_uint_t value);
-static const ngx_str_t *ngx_http_markdown_log_verbosity_name(ngx_uint_t value);
-static const ngx_str_t *ngx_http_markdown_compression_name(
-    ngx_http_markdown_compression_type_e compression_type);
-static const ngx_str_t *ngx_http_markdown_enabled_source_name(ngx_uint_t value);
-static ngx_uint_t ngx_http_markdown_is_ascii_space(u_char ch);
-static ngx_int_t ngx_http_markdown_parse_filter_flag(ngx_str_t *value, ngx_flag_t *enabled);
-ngx_flag_t ngx_http_markdown_is_enabled(ngx_http_request_t *r,
-    ngx_http_markdown_conf_t *conf);
-static void ngx_http_markdown_log_merged_conf(ngx_conf_t *cf, ngx_http_markdown_conf_t *conf);
-static ngx_int_t ngx_http_markdown_forward_headers(ngx_http_request_t *r,
-    ngx_http_markdown_ctx_t *ctx);
-static ngx_int_t ngx_http_markdown_send_buffered_original_response(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx);
-static ngx_int_t ngx_http_markdown_fail_open_with_buffered_prefix(
-    ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx, ngx_chain_t *remaining);
 #if !(NGX_HTTP_HEADERS)
 static ngx_table_elt_t *ngx_http_markdown_find_request_header(ngx_http_request_t *r,
     const ngx_str_t *name);
@@ -120,350 +63,9 @@ static ngx_table_elt_t *ngx_http_markdown_find_request_header(ngx_http_request_t
 static ngx_http_output_header_filter_pt ngx_http_next_header_filter;
 static ngx_http_output_body_filter_pt ngx_http_next_body_filter;
 
-/*
- * Module directives
- *
- * These directives control the behavior of the Markdown filter.
- * Each directive includes validation and clear error messages.
- */
-static ngx_command_t ngx_http_markdown_filter_commands[] = {
-    /*
-     * markdown_filter on|off|$variable
-     * 
-     * Enables or disables Markdown conversion for this context.
-     * Also supports per-request toggle via nginx variables/complex values.
-     * Default: off
-     * Context: http, server, location
-     * 
-     * Example:
-     *   location /api {
-     *       markdown_filter on;
-     *   }
-     */
-    {
-        ngx_string("markdown_filter"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_http_markdown_filter,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    /*
-     * markdown_max_size <size>
-     * 
-     * Maximum response size to attempt conversion (e.g., 10m, 5k).
-     * Responses larger than this will not be converted.
-     * Default: 10m (10 megabytes)
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_max_size 5m;
-     */
-    {
-        ngx_string("markdown_max_size"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_conf_set_size_slot,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        offsetof(ngx_http_markdown_conf_t, max_size),
-        NULL
-    },
-    
-    /*
-     * markdown_timeout <time>
-     * 
-     * Maximum time to spend on conversion (e.g., 5s, 1000ms).
-     * Conversions exceeding this timeout will be aborted.
-     * Default: 5s (5 seconds)
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_timeout 3s;
-     */
-    {
-        ngx_string("markdown_timeout"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_conf_set_msec_slot,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        offsetof(ngx_http_markdown_conf_t, timeout),
-        NULL
-    },
-    
-    /*
-     * markdown_on_error pass|reject
-     * 
-     * Failure strategy when conversion fails:
-     * - pass: Return original HTML (fail-open, default)
-     * - reject: Return 502 Bad Gateway (fail-closed)
-     * Default: pass
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_on_error reject;
-     */
-    {
-        ngx_string("markdown_on_error"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_http_markdown_on_error,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    /*
-     * markdown_flavor commonmark|gfm
-     * 
-     * Markdown flavor to generate:
-     * - commonmark: CommonMark specification (default)
-     * - gfm: GitHub Flavored Markdown (includes tables, strikethrough)
-     * Default: commonmark
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_flavor gfm;
-     */
-    {
-        ngx_string("markdown_flavor"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_http_markdown_flavor,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    /*
-     * markdown_token_estimate on|off
-     * 
-     * Include X-Markdown-Tokens header with estimated token count.
-     * Useful for AI agents to manage context windows.
-     * Default: off
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_token_estimate on;
-     */
-    {
-        ngx_string("markdown_token_estimate"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_FLAG,
-        ngx_conf_set_flag_slot,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        offsetof(ngx_http_markdown_conf_t, token_estimate),
-        NULL
-    },
-    
-    /*
-     * markdown_front_matter on|off
-     * 
-     * Include YAML front matter with metadata (title, description, etc.).
-     * Default: off
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_front_matter on;
-     */
-    {
-        ngx_string("markdown_front_matter"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_FLAG,
-        ngx_conf_set_flag_slot,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        offsetof(ngx_http_markdown_conf_t, front_matter),
-        NULL
-    },
-    
-    /*
-     * markdown_on_wildcard on|off
-     * 
-     * Convert when Accept header contains wildcards (star/slash-star or text slash star).
-     * Default: off (only convert on explicit text/markdown)
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_on_wildcard on;
-     */
-    {
-        ngx_string("markdown_on_wildcard"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_FLAG,
-        ngx_conf_set_flag_slot,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        offsetof(ngx_http_markdown_conf_t, on_wildcard),
-        NULL
-    },
-    
-    /*
-     * markdown_auth_policy allow|deny
-     * 
-     * Policy for converting authenticated requests:
-     * - allow: Convert authenticated requests (default)
-     * - deny: Skip conversion for authenticated requests
-     * Default: allow
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_auth_policy deny;
-     */
-    {
-        ngx_string("markdown_auth_policy"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_http_markdown_auth_policy,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    /*
-     * markdown_auth_cookies <pattern> [<pattern> ...]
-     * 
-     * Cookie name patterns to identify authenticated requests.
-     * Supports exact match, prefix match (pattern*), and wildcards.
-     * Default: none (only Authorization header detection)
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_auth_cookies session* auth_token PHPSESSID;
-     */
-    {
-        ngx_string("markdown_auth_cookies"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_1MORE,
-        ngx_http_markdown_auth_cookies,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    /*
-     * markdown_etag on|off
-     * 
-     * Generate ETag header for Markdown variants.
-     * ETags are computed from the Markdown output for proper caching.
-     * Default: on
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_etag off;
-     */
-    {
-        ngx_string("markdown_etag"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_FLAG,
-        ngx_conf_set_flag_slot,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        offsetof(ngx_http_markdown_conf_t, generate_etag),
-        NULL
-    },
-    
-    /*
-     * markdown_conditional_requests full_support|if_modified_since_only|disabled
-     * 
-     * Conditional request support mode:
-     * - full_support: Support If-None-Match and If-Modified-Since (default)
-     * - if_modified_since_only: Only support If-Modified-Since (performance)
-     * - disabled: No conditional request support for Markdown variants
-     * Default: full_support
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_conditional_requests if_modified_since_only;
-     */
-    {
-        ngx_string("markdown_conditional_requests"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_http_markdown_conditional_requests,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
+extern ngx_module_t ngx_http_markdown_filter_module;
 
-    /*
-     * markdown_log_verbosity error|warn|info|debug
-     *
-     * Module-local verbosity filter for module-generated logs.
-     * NGINX's global error_log level still applies.
-     * Default: info
-     * Context: http, server, location
-     *
-     * Example:
-     *   markdown_log_verbosity warn;
-     */
-    {
-        ngx_string("markdown_log_verbosity"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_TAKE1,
-        ngx_http_markdown_log_verbosity,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    /*
-     * markdown_buffer_chunked on|off
-     * 
-     * Buffer and convert chunked Transfer-Encoding responses.
-     * When off, chunked responses are passed through without conversion.
-     * Default: on
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_buffer_chunked off;
-     */
-    {
-        ngx_string("markdown_buffer_chunked"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_FLAG,
-        ngx_conf_set_flag_slot,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        offsetof(ngx_http_markdown_conf_t, buffer_chunked),
-        NULL
-    },
-    
-    /*
-     * markdown_stream_types <type> [<type> ...]
-     * 
-     * Content types to exclude from conversion (streaming responses).
-     * These content types will never be converted, even if eligible.
-     * Default: none (no exclusions)
-     * Context: http, server, location
-     * 
-     * Example:
-     *   markdown_stream_types text/event-stream application/x-ndjson;
-     */
-    {
-        ngx_string("markdown_stream_types"),
-        NGX_HTTP_MAIN_CONF|NGX_HTTP_SRV_CONF|NGX_HTTP_LOC_CONF|NGX_CONF_1MORE,
-        ngx_http_markdown_stream_types,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    /*
-     * markdown_metrics
-     * 
-     * Enables metrics exposure endpoint at this location.
-     * Returns metrics in plain text or JSON format based on Accept header.
-     * Access is restricted to localhost (127.0.0.1, ::1) by default for security.
-     * Default: off
-     * Context: location only
-     * 
-     * Example:
-     *   location /markdown-metrics {
-     *       markdown_metrics;
-     *   }
-     * 
-     * Response formats:
-     * - Plain text (default): Accept: text/plain or no Accept header
-     * - JSON: Accept: application/json
-     * 
-     * Security: Only accessible from localhost by default.
-     * Use allow/deny directives to customize access control.
-     */
-    {
-        ngx_string("markdown_metrics"),
-        NGX_HTTP_LOC_CONF|NGX_CONF_NOARGS,
-        ngx_http_markdown_metrics_directive,
-        NGX_HTTP_LOC_CONF_OFFSET,
-        0,
-        NULL
-    },
-    
-    ngx_null_command
-};
+#include "ngx_http_markdown_config_impl.h"
 
 /*
  * Module context
@@ -473,7 +75,7 @@ static ngx_command_t ngx_http_markdown_filter_commands[] = {
 static ngx_http_module_t ngx_http_markdown_filter_module_ctx = {
     NULL,                                   /* preconfiguration */
     ngx_http_markdown_filter_init,          /* postconfiguration */
-    NULL,                                   /* create main configuration */
+    ngx_http_markdown_create_main_conf,     /* create main configuration */
     NULL,                                   /* init main configuration */
     NULL,                                   /* create server configuration */
     NULL,                                   /* merge server configuration */
@@ -499,2653 +101,11 @@ ngx_module_t ngx_http_markdown_filter_module = {
     NGX_MODULE_V1_PADDING
 };
 
-/*
- * Create location configuration
- *
- * Allocates and initializes configuration structure with default values.
- * All values are set to NGX_CONF_UNSET* to enable proper inheritance.
- */
-static void *
-ngx_http_markdown_create_conf(ngx_conf_t *cf)
-{
-    ngx_http_markdown_conf_t *conf;
-    
-    conf = ngx_pcalloc(cf->pool, sizeof(ngx_http_markdown_conf_t));
-    if (conf == NULL) {
-        return NULL;
-    }
-    
-    /* Set unset values (NGX_CONF_UNSET*) for proper inheritance */
-    conf->enabled = NGX_CONF_UNSET;
-    conf->enabled_source = NGX_HTTP_MARKDOWN_ENABLED_UNSET;
-    conf->enabled_complex = NULL;
-    conf->max_size = NGX_CONF_UNSET_SIZE;
-    conf->timeout = NGX_CONF_UNSET_MSEC;
-    conf->on_error = NGX_CONF_UNSET_UINT;
-    conf->flavor = NGX_CONF_UNSET_UINT;
-    conf->token_estimate = NGX_CONF_UNSET;
-    conf->front_matter = NGX_CONF_UNSET;
-    conf->on_wildcard = NGX_CONF_UNSET;
-    conf->auth_policy = NGX_CONF_UNSET_UINT;
-    conf->auth_cookies = NGX_CONF_UNSET_PTR;
-    conf->generate_etag = NGX_CONF_UNSET;
-    conf->conditional_requests = NGX_CONF_UNSET_UINT;
-    conf->log_verbosity = NGX_CONF_UNSET_UINT;
-    conf->buffer_chunked = NGX_CONF_UNSET;
-    conf->stream_types = NGX_CONF_UNSET_PTR;
-    conf->auto_decompress = NGX_CONF_UNSET;
-    
-    return conf;
-}
+#include "ngx_http_markdown_lifecycle_impl.h"
 
-/*
- * Merge location configuration
- *
- * Implements configuration inheritance: location > server > http.
- * Child configuration values override parent values if set.
- * Unset child values inherit from parent.
- *
- * Default values (applied when both parent and child are unset):
- * - enabled: 0 (off)
- * - max_size: 10MB (10 * 1024 * 1024 bytes)
- * - timeout: 5000ms (5 seconds)
- * - on_error: NGX_HTTP_MARKDOWN_ON_ERROR_PASS (fail-open)
- * - flavor: NGX_HTTP_MARKDOWN_FLAVOR_COMMONMARK
- * - token_estimate: 0 (off)
- * - front_matter: 0 (off)
- * - on_wildcard: 0 (off)
- * - auth_policy: NGX_HTTP_MARKDOWN_AUTH_POLICY_ALLOW
- * - auth_cookies: NULL (no patterns)
- * - generate_etag: 1 (on)
- * - conditional_requests: NGX_HTTP_MARKDOWN_CONDITIONAL_FULL_SUPPORT
- * - log_verbosity: NGX_HTTP_MARKDOWN_LOG_INFO
- * - buffer_chunked: 1 (on)
- * - stream_types: NULL (no exclusions)
- * - auto_decompress: 1 (on)
- */
-static char *
-ngx_http_markdown_merge_conf(ngx_conf_t *cf, void *parent, void *child)
-{
-    ngx_http_markdown_conf_t *prev = parent;
-    ngx_http_markdown_conf_t *conf = child;
-    
-    /*
-     * Merge markdown_filter with explicit source tracking.
-     *
-     * Priority:
-     * 1. Child explicit value (on/off or complex expression)
-     * 2. Parent inherited value
-     * 3. Default off
-     */
-    if (conf->enabled_source == NGX_HTTP_MARKDOWN_ENABLED_UNSET) {
-        if (prev->enabled_source == NGX_HTTP_MARKDOWN_ENABLED_UNSET) {
-            conf->enabled_source = NGX_HTTP_MARKDOWN_ENABLED_STATIC;
-            conf->enabled = 0;
-            conf->enabled_complex = NULL;
-        } else {
-            conf->enabled_source = prev->enabled_source;
-            conf->enabled = prev->enabled;
-            conf->enabled_complex = prev->enabled_complex;
-        }
-    } else if (conf->enabled_source == NGX_HTTP_MARKDOWN_ENABLED_STATIC) {
-        conf->enabled_complex = NULL;
-    }
+#include "ngx_http_markdown_request_impl.h"
 
-    /* Merge flag and scalar configuration values */
-    ngx_conf_merge_size_value(conf->max_size, prev->max_size, 10 * 1024 * 1024); /* 10MB */
-    ngx_conf_merge_msec_value(conf->timeout, prev->timeout, 5000); /* 5 seconds */
-    ngx_conf_merge_uint_value(conf->on_error, prev->on_error, 
-                              NGX_HTTP_MARKDOWN_ON_ERROR_PASS);
-    ngx_conf_merge_uint_value(conf->flavor, prev->flavor, 
-                              NGX_HTTP_MARKDOWN_FLAVOR_COMMONMARK);
-    ngx_conf_merge_value(conf->token_estimate, prev->token_estimate, 0);
-    ngx_conf_merge_value(conf->front_matter, prev->front_matter, 0);
-    ngx_conf_merge_value(conf->on_wildcard, prev->on_wildcard, 0);
-    ngx_conf_merge_uint_value(conf->auth_policy, prev->auth_policy, 
-                              NGX_HTTP_MARKDOWN_AUTH_POLICY_ALLOW);
-    ngx_conf_merge_value(conf->generate_etag, prev->generate_etag, 1);
-    ngx_conf_merge_uint_value(conf->conditional_requests, prev->conditional_requests,
-                              NGX_HTTP_MARKDOWN_CONDITIONAL_FULL_SUPPORT);
-    ngx_conf_merge_uint_value(conf->log_verbosity, prev->log_verbosity,
-                              NGX_HTTP_MARKDOWN_LOG_INFO);
-    ngx_conf_merge_value(conf->buffer_chunked, prev->buffer_chunked, 1);
-    ngx_conf_merge_value(conf->auto_decompress, prev->auto_decompress, 1);
-    
-    /* Merge pointer configuration values (arrays) */
-    ngx_conf_merge_ptr_value(conf->auth_cookies, prev->auth_cookies, NULL);
-    ngx_conf_merge_ptr_value(conf->stream_types, prev->stream_types, NULL);
-
-    /* Startup/reload-time configuration snapshot (FR-12.7) */
-    ngx_http_markdown_log_merged_conf(cf, conf);
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_filter
- *
- * Supported values:
- * - on | off          (static configuration)
- * - $variable         (dynamic per-request switch)
- * - complex value containing variables
- */
-static char *
-ngx_http_markdown_filter(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t          *mcf = conf;
-    ngx_http_compile_complex_value_t   ccv;
-    ngx_http_complex_value_t          *complex_value;
-    ngx_str_t                         *value;
-    u_char                            *var_marker;
-
-    (void) cmd;
-
-    if (mcf->enabled_source != NGX_HTTP_MARKDOWN_ENABLED_UNSET) {
-        return "is duplicate";
-    }
-
-    value = cf->args->elts;
-
-    if (value[1].len == 2
-        && ngx_strcasecmp(value[1].data, (u_char *) "on") == 0)
-    {
-        mcf->enabled = 1;
-        mcf->enabled_source = NGX_HTTP_MARKDOWN_ENABLED_STATIC;
-        mcf->enabled_complex = NULL;
-        return NGX_CONF_OK;
-    }
-
-    if (value[1].len == 3
-        && ngx_strcasecmp(value[1].data, (u_char *) "off") == 0)
-    {
-        mcf->enabled = 0;
-        mcf->enabled_source = NGX_HTTP_MARKDOWN_ENABLED_STATIC;
-        mcf->enabled_complex = NULL;
-        return NGX_CONF_OK;
-    }
-
-    var_marker = ngx_strlchr(value[1].data, value[1].data + value[1].len, '$');
-    if (var_marker == NULL) {
-        ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                           "invalid value \"%V\" in \"markdown_filter\" directive, "
-                           "it must be \"on\", \"off\", or contain a variable",
-                           &value[1]);
-        return NGX_CONF_ERROR;
-    }
-
-    complex_value = ngx_palloc(cf->pool, sizeof(ngx_http_complex_value_t));
-    if (complex_value == NULL) {
-        return NGX_CONF_ERROR;
-    }
-
-    ngx_memzero(&ccv, sizeof(ngx_http_compile_complex_value_t));
-    ccv.cf = cf;
-    ccv.value = &value[1];
-    ccv.complex_value = complex_value;
-
-    if (ngx_http_compile_complex_value(&ccv) != NGX_OK) {
-        return NGX_CONF_ERROR;
-    }
-
-    mcf->enabled = 0;
-    mcf->enabled_source = NGX_HTTP_MARKDOWN_ENABLED_COMPLEX;
-    mcf->enabled_complex = complex_value;
-
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_on_error
- *
- * Parses and validates the markdown_on_error directive value.
- * Valid values: "pass" (fail-open) or "reject" (fail-closed)
- *
- * @param cf    Configuration structure
- * @param cmd   Command structure
- * @param conf  Module configuration
- * @return      NGX_CONF_OK on success, NGX_CONF_ERROR on error
- */
-static char *
-ngx_http_markdown_on_error(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t *mcf = conf;
-    ngx_str_t                *value;
-    
-    /* Get directive arguments */
-    value = cf->args->elts;
-    
-    /* Check if already set in this context */
-    if (mcf->on_error != NGX_CONF_UNSET_UINT) {
-        return "is duplicate";
-    }
-    
-    /* Parse and validate value */
-    if (ngx_strcmp(value[1].data, "pass") == 0) {
-        mcf->on_error = NGX_HTTP_MARKDOWN_ON_ERROR_PASS;
-    } else if (ngx_strcmp(value[1].data, "reject") == 0) {
-        mcf->on_error = NGX_HTTP_MARKDOWN_ON_ERROR_REJECT;
-    } else {
-        ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                          "invalid value \"%V\" in \"%V\" directive, "
-                          "it must be \"pass\" or \"reject\"",
-                          &value[1], &cmd->name);
-        return NGX_CONF_ERROR;
-    }
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_flavor
- *
- * Parses and validates the markdown_flavor directive value.
- * Valid values: "commonmark" or "gfm" (GitHub Flavored Markdown)
- *
- * @param cf    Configuration structure
- * @param cmd   Command structure
- * @param conf  Module configuration
- * @return      NGX_CONF_OK on success, NGX_CONF_ERROR on error
- */
-static char *
-ngx_http_markdown_flavor(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t *mcf = conf;
-    ngx_str_t                *value;
-    
-    /* Get directive arguments */
-    value = cf->args->elts;
-    
-    /* Check if already set in this context */
-    if (mcf->flavor != NGX_CONF_UNSET_UINT) {
-        return "is duplicate";
-    }
-    
-    /* Parse and validate value */
-    if (ngx_strcmp(value[1].data, "commonmark") == 0) {
-        mcf->flavor = NGX_HTTP_MARKDOWN_FLAVOR_COMMONMARK;
-    } else if (ngx_strcmp(value[1].data, "gfm") == 0) {
-        mcf->flavor = NGX_HTTP_MARKDOWN_FLAVOR_GFM;
-    } else {
-        ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                          "invalid value \"%V\" in \"%V\" directive, "
-                          "it must be \"commonmark\" or \"gfm\"",
-                          &value[1], &cmd->name);
-        return NGX_CONF_ERROR;
-    }
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_auth_policy
- *
- * Parses and validates the markdown_auth_policy directive value.
- * Valid values: "allow" or "deny"
- *
- * @param cf    Configuration structure
- * @param cmd   Command structure
- * @param conf  Module configuration
- * @return      NGX_CONF_OK on success, NGX_CONF_ERROR on error
- */
-static char *
-ngx_http_markdown_auth_policy(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t *mcf = conf;
-    ngx_str_t                *value;
-    
-    /* Get directive arguments */
-    value = cf->args->elts;
-    
-    /* Check if already set in this context */
-    if (mcf->auth_policy != NGX_CONF_UNSET_UINT) {
-        return "is duplicate";
-    }
-    
-    /* Parse and validate value */
-    if (ngx_strcmp(value[1].data, "allow") == 0) {
-        mcf->auth_policy = NGX_HTTP_MARKDOWN_AUTH_POLICY_ALLOW;
-    } else if (ngx_strcmp(value[1].data, "deny") == 0) {
-        mcf->auth_policy = NGX_HTTP_MARKDOWN_AUTH_POLICY_DENY;
-    } else {
-        ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                          "invalid value \"%V\" in \"%V\" directive, "
-                          "it must be \"allow\" or \"deny\"",
-                          &value[1], &cmd->name);
-        return NGX_CONF_ERROR;
-    }
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_auth_cookies
- *
- * Parses cookie name patterns for authentication detection.
- * Supports multiple patterns (exact match, prefix match with *).
- *
- * Examples:
- *   markdown_auth_cookies session* auth_token;
- *   markdown_auth_cookies PHPSESSID wordpress_logged_in_*;
- *
- * @param cf    Configuration structure
- * @param cmd   Command structure
- * @param conf  Module configuration
- * @return      NGX_CONF_OK on success, NGX_CONF_ERROR on error
- */
-static char *
-ngx_http_markdown_auth_cookies(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t *mcf = conf;
-    ngx_str_t                *value;
-    ngx_str_t                *pattern;
-    ngx_uint_t                i;
-    
-    /* Get directive arguments */
-    value = cf->args->elts;
-    
-    /* Check if already set in this context */
-    if (mcf->auth_cookies != NGX_CONF_UNSET_PTR) {
-        return "is duplicate";
-    }
-    
-    /* Create array to store patterns */
-    mcf->auth_cookies = ngx_array_create(cf->pool, cf->args->nelts - 1, sizeof(ngx_str_t));
-    if (mcf->auth_cookies == NULL) {
-        return NGX_CONF_ERROR;
-    }
-    
-    /* Add each pattern to the array */
-    for (i = 1; i < cf->args->nelts; i++) {
-        /* Validate pattern is not empty */
-        if (value[i].len == 0) {
-            ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                              "empty cookie pattern in \"%V\" directive",
-                              &cmd->name);
-            return NGX_CONF_ERROR;
-        }
-        
-        /* Allocate and copy pattern */
-        pattern = ngx_array_push(mcf->auth_cookies);
-        if (pattern == NULL) {
-            return NGX_CONF_ERROR;
-        }
-        
-        *pattern = value[i];
-        
-        ngx_conf_log_error(NGX_LOG_DEBUG, cf, 0,
-                          "markdown_auth_cookies: added pattern \"%V\"",
-                          pattern);
-    }
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_conditional_requests
- *
- * Parses and validates the markdown_conditional_requests directive value.
- * Valid values: "full_support", "if_modified_since_only", or "disabled"
- *
- * @param cf    Configuration structure
- * @param cmd   Command structure
- * @param conf  Module configuration
- * @return      NGX_CONF_OK on success, NGX_CONF_ERROR on error
- */
-static char *
-ngx_http_markdown_conditional_requests(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t *mcf = conf;
-    ngx_str_t                *value;
-    
-    /* Get directive arguments */
-    value = cf->args->elts;
-    
-    /* Check if already set in this context */
-    if (mcf->conditional_requests != NGX_CONF_UNSET_UINT) {
-        return "is duplicate";
-    }
-    
-    /* Parse and validate value */
-    if (ngx_strcmp(value[1].data, "full_support") == 0) {
-        mcf->conditional_requests = NGX_HTTP_MARKDOWN_CONDITIONAL_FULL_SUPPORT;
-    } else if (ngx_strcmp(value[1].data, "if_modified_since_only") == 0) {
-        mcf->conditional_requests = NGX_HTTP_MARKDOWN_CONDITIONAL_IF_MODIFIED_SINCE;
-    } else if (ngx_strcmp(value[1].data, "disabled") == 0) {
-        mcf->conditional_requests = NGX_HTTP_MARKDOWN_CONDITIONAL_DISABLED;
-    } else {
-        ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                          "invalid value \"%V\" in \"%V\" directive, "
-                          "it must be \"full_support\", \"if_modified_since_only\", or \"disabled\"",
-                          &value[1], &cmd->name);
-        return NGX_CONF_ERROR;
-    }
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_log_verbosity
- *
- * Valid values: "error", "warn", "info", "debug"
- */
-static char *
-ngx_http_markdown_log_verbosity(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t *mcf = conf;
-    ngx_str_t                *value;
-
-    value = cf->args->elts;
-
-    if (mcf->log_verbosity != NGX_CONF_UNSET_UINT) {
-        return "is duplicate";
-    }
-
-    if (ngx_strcmp(value[1].data, "error") == 0) {
-        mcf->log_verbosity = NGX_HTTP_MARKDOWN_LOG_ERROR;
-    } else if (ngx_strcmp(value[1].data, "warn") == 0) {
-        mcf->log_verbosity = NGX_HTTP_MARKDOWN_LOG_WARN;
-    } else if (ngx_strcmp(value[1].data, "info") == 0) {
-        mcf->log_verbosity = NGX_HTTP_MARKDOWN_LOG_INFO;
-    } else if (ngx_strcmp(value[1].data, "debug") == 0) {
-        mcf->log_verbosity = NGX_HTTP_MARKDOWN_LOG_DEBUG;
-    } else {
-        ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                          "invalid value \"%V\" in \"%V\" directive, "
-                          "it must be \"error\", \"warn\", \"info\", or \"debug\"",
-                          &value[1], &cmd->name);
-        return NGX_CONF_ERROR;
-    }
-
-    return NGX_CONF_OK;
-}
-
-static ngx_uint_t
-ngx_http_markdown_log_verbosity_to_ngx_level(ngx_uint_t verbosity)
-{
-    switch (verbosity) {
-        case NGX_HTTP_MARKDOWN_LOG_ERROR:
-            return NGX_LOG_ERR;
-        case NGX_HTTP_MARKDOWN_LOG_WARN:
-            return NGX_LOG_WARN;
-        case NGX_HTTP_MARKDOWN_LOG_DEBUG:
-            return NGX_LOG_DEBUG;
-        case NGX_HTTP_MARKDOWN_LOG_INFO:
-        default:
-            return NGX_LOG_INFO;
-    }
-}
-
-static const ngx_str_t *
-ngx_http_markdown_on_error_name(ngx_uint_t value)
-{
-    static ngx_str_t  pass = ngx_string("pass");
-    static ngx_str_t  reject = ngx_string("reject");
-    static ngx_str_t  unknown = ngx_string("unknown");
-
-    switch (value) {
-        case NGX_HTTP_MARKDOWN_ON_ERROR_PASS:
-            return &pass;
-        case NGX_HTTP_MARKDOWN_ON_ERROR_REJECT:
-            return &reject;
-        default:
-            return &unknown;
-    }
-}
-
-static const ngx_str_t *
-ngx_http_markdown_flavor_name(ngx_uint_t value)
-{
-    static ngx_str_t  commonmark = ngx_string("commonmark");
-    static ngx_str_t  gfm = ngx_string("gfm");
-    static ngx_str_t  unknown = ngx_string("unknown");
-
-    switch (value) {
-        case NGX_HTTP_MARKDOWN_FLAVOR_COMMONMARK:
-            return &commonmark;
-        case NGX_HTTP_MARKDOWN_FLAVOR_GFM:
-            return &gfm;
-        default:
-            return &unknown;
-    }
-}
-
-static const ngx_str_t *
-ngx_http_markdown_auth_policy_name(ngx_uint_t value)
-{
-    static ngx_str_t  allow = ngx_string("allow");
-    static ngx_str_t  deny = ngx_string("deny");
-    static ngx_str_t  unknown = ngx_string("unknown");
-
-    switch (value) {
-        case NGX_HTTP_MARKDOWN_AUTH_POLICY_ALLOW:
-            return &allow;
-        case NGX_HTTP_MARKDOWN_AUTH_POLICY_DENY:
-            return &deny;
-        default:
-            return &unknown;
-    }
-}
-
-static const ngx_str_t *
-ngx_http_markdown_conditional_requests_name(ngx_uint_t value)
-{
-    static ngx_str_t  full_support = ngx_string("full_support");
-    static ngx_str_t  if_modified_since_only = ngx_string("if_modified_since_only");
-    static ngx_str_t  disabled = ngx_string("disabled");
-    static ngx_str_t  unknown = ngx_string("unknown");
-
-    switch (value) {
-        case NGX_HTTP_MARKDOWN_CONDITIONAL_FULL_SUPPORT:
-            return &full_support;
-        case NGX_HTTP_MARKDOWN_CONDITIONAL_IF_MODIFIED_SINCE:
-            return &if_modified_since_only;
-        case NGX_HTTP_MARKDOWN_CONDITIONAL_DISABLED:
-            return &disabled;
-        default:
-            return &unknown;
-    }
-}
-
-static const ngx_str_t *
-ngx_http_markdown_log_verbosity_name(ngx_uint_t value)
-{
-    static ngx_str_t  error = ngx_string("error");
-    static ngx_str_t  warn = ngx_string("warn");
-    static ngx_str_t  info = ngx_string("info");
-    static ngx_str_t  debug = ngx_string("debug");
-    static ngx_str_t  unknown = ngx_string("unknown");
-
-    switch (value) {
-        case NGX_HTTP_MARKDOWN_LOG_ERROR:
-            return &error;
-        case NGX_HTTP_MARKDOWN_LOG_WARN:
-            return &warn;
-        case NGX_HTTP_MARKDOWN_LOG_INFO:
-            return &info;
-        case NGX_HTTP_MARKDOWN_LOG_DEBUG:
-            return &debug;
-        default:
-            return &unknown;
-    }
-}
-
-static const ngx_str_t *
-ngx_http_markdown_compression_name(ngx_http_markdown_compression_type_e compression_type)
-{
-    static ngx_str_t  gzip = ngx_string("gzip");
-    static ngx_str_t  deflate = ngx_string("deflate");
-    static ngx_str_t  brotli = ngx_string("brotli");
-    static ngx_str_t  unknown = ngx_string("unknown");
-    static ngx_str_t  invalid = ngx_string("invalid");
-
-    switch (compression_type) {
-        case NGX_HTTP_MARKDOWN_COMPRESSION_GZIP:
-            return &gzip;
-        case NGX_HTTP_MARKDOWN_COMPRESSION_DEFLATE:
-            return &deflate;
-        case NGX_HTTP_MARKDOWN_COMPRESSION_BROTLI:
-            return &brotli;
-        case NGX_HTTP_MARKDOWN_COMPRESSION_UNKNOWN:
-            return &unknown;
-        default:
-            return &invalid;
-    }
-}
-
-static const ngx_str_t *
-ngx_http_markdown_enabled_source_name(ngx_uint_t value)
-{
-    static ngx_str_t unset = ngx_string("unset");
-    static ngx_str_t static_value = ngx_string("static");
-    static ngx_str_t complex = ngx_string("complex");
-    static ngx_str_t unknown = ngx_string("unknown");
-
-    switch (value) {
-        case NGX_HTTP_MARKDOWN_ENABLED_UNSET:
-            return &unset;
-        case NGX_HTTP_MARKDOWN_ENABLED_STATIC:
-            return &static_value;
-        case NGX_HTTP_MARKDOWN_ENABLED_COMPLEX:
-            return &complex;
-        default:
-            return &unknown;
-    }
-}
-
-static ngx_uint_t
-ngx_http_markdown_is_ascii_space(u_char ch)
-{
-    return (ch == ' ' || ch == '\t' || ch == '\r'
-            || ch == '\n' || ch == '\f' || ch == '\v');
-}
-
-static u_char ngx_http_markdown_flag_on[] = "on";
-static u_char ngx_http_markdown_flag_off[] = "off";
-static u_char ngx_http_markdown_flag_yes[] = "yes";
-static u_char ngx_http_markdown_flag_no[] = "no";
-static u_char ngx_http_markdown_flag_true[] = "true";
-static u_char ngx_http_markdown_flag_false[] = "false";
-
-static ngx_int_t
-ngx_http_markdown_parse_filter_flag(ngx_str_t *value, ngx_flag_t *enabled)
-{
-    ngx_str_t  normalized;
-    u_char    *start;
-    u_char    *end;
-
-    if (value == NULL || enabled == NULL) {
-        return NGX_ERROR;
-    }
-
-    /*
-     * Normalize by trimming surrounding ASCII whitespace.
-     * This keeps per-request variable parsing tolerant to values such as
-     * " on " or "\t1\n" without allocating new buffers.
-     */
-    start = value->data;
-    end = value->data + value->len;
-
-    while (start < end && ngx_http_markdown_is_ascii_space(*start)) {
-        start++;
-    }
-
-    while (end > start && ngx_http_markdown_is_ascii_space(*(end - 1))) {
-        end--;
-    }
-
-    normalized.data = start;
-    normalized.len = (size_t) (end - start);
-    value = &normalized;
-
-    if (value->len == 0) {
-        *enabled = 0;
-        return NGX_OK;
-    }
-
-    if (value->len == 1) {
-        if (value->data[0] == '1') {
-            *enabled = 1;
-            return NGX_OK;
-        }
-
-        if (value->data[0] == '0') {
-            *enabled = 0;
-            return NGX_OK;
-        }
-    }
-
-    if (value->len == 2
-        && ngx_strncasecmp(value->data, ngx_http_markdown_flag_on, 2) == 0)
-    {
-        *enabled = 1;
-        return NGX_OK;
-    }
-
-    if (value->len == 3
-        && ngx_strncasecmp(value->data, ngx_http_markdown_flag_off, 3) == 0)
-    {
-        *enabled = 0;
-        return NGX_OK;
-    }
-
-    if (value->len == 3
-        && ngx_strncasecmp(value->data, ngx_http_markdown_flag_yes, 3) == 0)
-    {
-        *enabled = 1;
-        return NGX_OK;
-    }
-
-    if (value->len == 2
-        && ngx_strncasecmp(value->data, ngx_http_markdown_flag_no, 2) == 0)
-    {
-        *enabled = 0;
-        return NGX_OK;
-    }
-
-    if (value->len == 4
-        && ngx_strncasecmp(value->data, ngx_http_markdown_flag_true, 4) == 0)
-    {
-        *enabled = 1;
-        return NGX_OK;
-    }
-
-    if (value->len == 5
-        && ngx_strncasecmp(value->data, ngx_http_markdown_flag_false, 5) == 0)
-    {
-        *enabled = 0;
-        return NGX_OK;
-    }
-
-    return NGX_ERROR;
-}
-
-ngx_flag_t
-ngx_http_markdown_is_enabled(ngx_http_request_t *r, ngx_http_markdown_conf_t *conf)
-{
-    ngx_str_t    evaluated;
-    ngx_flag_t   enabled;
-    ngx_int_t    rc;
-
-    if (conf == NULL) {
-        return 0;
-    }
-
-    if (conf->enabled_source != NGX_HTTP_MARKDOWN_ENABLED_COMPLEX
-        || conf->enabled_complex == NULL)
-    {
-        return conf->enabled;
-    }
-
-    if (r == NULL) {
-        return 0;
-    }
-
-    if (ngx_http_complex_value(r, conf->enabled_complex, &evaluated) != NGX_OK) {
-        ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                      "markdown filter: failed to evaluate markdown_filter variable");
-        return 0;
-    }
-
-    rc = ngx_http_markdown_parse_filter_flag(&evaluated, &enabled);
-    if (rc != NGX_OK) {
-        ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                      "markdown filter: markdown_filter variable resolved to invalid value "
-                      "\"%V\", treating as off", &evaluated);
-        return 0;
-    }
-
-    return enabled;
-}
-
-static void
-ngx_http_markdown_log_merged_conf(ngx_conf_t *cf, ngx_http_markdown_conf_t *conf)
-{
-    ngx_uint_t log_level;
-    ngx_uint_t auth_cookie_count = (conf->auth_cookies != NULL) ? conf->auth_cookies->nelts : 0;
-    ngx_uint_t stream_type_count = (conf->stream_types != NULL) ? conf->stream_types->nelts : 0;
-
-    if (cf == NULL) {
-        return;
-    }
-
-    log_level = ngx_http_markdown_log_verbosity_to_ngx_level(conf->log_verbosity);
-
-    ngx_conf_log_error(log_level, cf, 0,
-                      "markdown filter config: enabled=%ui enabled_source=%V max_size=%uz timeout_ms=%M "
-                      "on_error=%V flavor=%V token_estimate=%ui front_matter=%ui "
-                      "on_wildcard=%ui auth_policy=%V auth_cookie_patterns=%ui "
-                      "etag=%ui conditional_requests=%V log_verbosity=%V "
-                      "buffer_chunked=%ui stream_types=%ui",
-                      (ngx_uint_t) conf->enabled,
-                      ngx_http_markdown_enabled_source_name(conf->enabled_source),
-                      conf->max_size,
-                      conf->timeout,
-                      ngx_http_markdown_on_error_name(conf->on_error),
-                      ngx_http_markdown_flavor_name(conf->flavor),
-                      (ngx_uint_t) conf->token_estimate,
-                      (ngx_uint_t) conf->front_matter,
-                      (ngx_uint_t) conf->on_wildcard,
-                      ngx_http_markdown_auth_policy_name(conf->auth_policy),
-                      auth_cookie_count,
-                      (ngx_uint_t) conf->generate_etag,
-                      ngx_http_markdown_conditional_requests_name(conf->conditional_requests),
-                      ngx_http_markdown_log_verbosity_name(conf->log_verbosity),
-                      (ngx_uint_t) conf->buffer_chunked,
-                      stream_type_count);
-}
-
-/*
- * Configuration directive handler: markdown_stream_types
- *
- * Parses content type patterns to exclude from conversion.
- * These are typically streaming content types that should never be buffered.
- *
- * Examples:
- *   markdown_stream_types text/event-stream;
- *   markdown_stream_types text/event-stream application/x-ndjson;
- *
- * @param cf    Configuration structure
- * @param cmd   Command structure
- * @param conf  Module configuration
- * @return      NGX_CONF_OK on success, NGX_CONF_ERROR on error
- */
-static char *
-ngx_http_markdown_stream_types(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_markdown_conf_t *mcf = conf;
-    ngx_str_t                *value;
-    ngx_str_t                *type;
-    ngx_uint_t                i;
-    
-    /* Get directive arguments */
-    value = cf->args->elts;
-    
-    /* Check if already set in this context */
-    if (mcf->stream_types != NGX_CONF_UNSET_PTR) {
-        return "is duplicate";
-    }
-    
-    /* Create array to store content types */
-    mcf->stream_types = ngx_array_create(cf->pool, cf->args->nelts - 1, sizeof(ngx_str_t));
-    if (mcf->stream_types == NULL) {
-        return NGX_CONF_ERROR;
-    }
-    
-    /* Add each content type to the array */
-    for (i = 1; i < cf->args->nelts; i++) {
-        /* Validate content type is not empty */
-        if (value[i].len == 0) {
-            ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                              "empty content type in \"%V\" directive",
-                              &cmd->name);
-            return NGX_CONF_ERROR;
-        }
-        
-        /* Basic validation: must contain a slash */
-        if (ngx_strchr(value[i].data, '/') == NULL) {
-            ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                              "invalid content type \"%V\" in \"%V\" directive, "
-                              "must be in format \"type/subtype\"",
-                              &value[i], &cmd->name);
-            return NGX_CONF_ERROR;
-        }
-        
-        /* Allocate and copy content type */
-        type = ngx_array_push(mcf->stream_types);
-        if (type == NULL) {
-            return NGX_CONF_ERROR;
-        }
-        
-        *type = value[i];
-        
-        ngx_conf_log_error(NGX_LOG_DEBUG, cf, 0,
-                          "markdown_stream_types: added type \"%V\"",
-                          type);
-    }
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Configuration directive handler: markdown_metrics
- *
- * Enables metrics exposure endpoint at this location.
- * The endpoint returns metrics in plain text or JSON format based on Accept header.
- * Access is restricted to localhost (127.0.0.1, ::1) by default for security.
- *
- * Example:
- *   location /markdown-metrics {
- *       markdown_metrics;
- *       # Optional: customize access control
- *       allow 10.0.0.0/8;
- *       deny all;
- *   }
- *
- * @param cf    Configuration structure
- * @param cmd   Command structure
- * @param conf  Module configuration (unused, directive takes no arguments)
- * @return      NGX_CONF_OK on success, NGX_CONF_ERROR on error
- *
- * Requirements: FR-13.5
- * Task: 22.4 Implement metrics exposure endpoint
- */
-static char *
-ngx_http_markdown_metrics_directive(ngx_conf_t *cf, ngx_command_t *cmd, void *conf)
-{
-    ngx_http_core_loc_conf_t  *clcf;
-    
-    /*
-     * Get the core location configuration for this location block.
-     * We need to set the handler for this location to our metrics handler.
-     */
-    clcf = ngx_http_conf_get_module_loc_conf(cf, ngx_http_core_module);
-    if (clcf == NULL) {
-        ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
-                          "failed to get core location configuration for \"%V\" directive",
-                          &cmd->name);
-        return NGX_CONF_ERROR;
-    }
-    
-    /*
-     * Set the handler for this location to our metrics handler.
-     * When a request matches this location, ngx_http_markdown_metrics_handler
-     * will be called to generate the metrics response.
-     */
-    clcf->handler = ngx_http_markdown_metrics_handler;
-    
-    ngx_conf_log_error(NGX_LOG_INFO, cf, 0,
-                      "markdown_metrics: endpoint enabled at this location");
-    
-    return NGX_CONF_OK;
-}
-
-/*
- * Module initialization
- *
- * Registers filter hooks in the NGINX filter chain.
- * This function is called during NGINX configuration phase to insert
- * our filters into the header and body filter chains.
- */
-static ngx_int_t
-ngx_http_markdown_filter_init(ngx_conf_t *cf)
-{
-    /*
-     * Insert into header filter chain
-     * 
-     * The header filter is called when response headers are being sent.
-     * We use it to check if conversion should be attempted and set up
-     * the conversion context.
-     */
-    ngx_http_next_header_filter = ngx_http_top_header_filter;
-    ngx_http_top_header_filter = ngx_http_markdown_header_filter;
-    
-    /*
-     * Insert into body filter chain
-     * 
-     * The body filter is called for each chunk of the response body.
-     * We use it to buffer the response and perform conversion when complete.
-     */
-    ngx_http_next_body_filter = ngx_http_top_body_filter;
-    ngx_http_top_body_filter = ngx_http_markdown_body_filter;
-    
-    return NGX_OK;
-}
-
-/*
- * Worker process initialization
- *
- * Called once per worker process when NGINX starts or reloads.
- * Initializes the Rust converter instance for this worker.
- *
- * NGINX uses a multi-process worker model where each worker process
- * handles requests independently. Each worker needs its own converter
- * instance because:
- * 1. The converter is not thread-safe across workers
- * 2. Each worker operates in its own memory space
- * 3. No shared state is needed between workers
- *
- * @param cycle  The NGINX cycle structure
- * @return       NGX_OK on success, NGX_ERROR on failure
- */
-static ngx_int_t
-ngx_http_markdown_init_worker(ngx_cycle_t *cycle)
-{
-    /*
-     * Initialize metrics structure.
-     * 
-     * All counters are initialized to zero. Since this is a global variable
-     * in the worker process, it will be automatically zero-initialized by
-     * the C runtime, but we explicitly zero it here for clarity.
-     * 
-     * Requirements: FR-09.8, FR-13.4
-     */
-    ngx_memzero(&ngx_http_markdown_metrics, sizeof(ngx_http_markdown_metrics_t));
-    
-    /*
-     * Create a new converter instance for this worker process.
-     * 
-     * The converter is stored in a global variable that is local to
-     * this worker process. Since NGINX uses a multi-process model
-     * (not multi-threaded), this global is safe and won't be shared
-     * across workers.
-     */
-    ngx_http_markdown_converter = markdown_converter_new();
-    
-    if (ngx_http_markdown_converter == NULL) {
-        /*
-         * Converter initialization failed.
-         * 
-         * This is a critical error because without a converter, the module
-         * cannot function. We log a critical-level message and return NGX_ERROR,
-         * which will cause NGINX to fail the worker initialization.
-         * 
-         * In production, this typically means:
-         * - Memory allocation failed (system out of memory)
-         * - Rust library initialization failed
-         * 
-         * The worker will not start, and NGINX will either:
-         * - Retry with another worker (if configured)
-         * - Fail to start entirely (if all workers fail)
-         * 
-         * Log level: NGX_LOG_CRIT (critical system failure)
-         * Requirements: FR-09.5, FR-09.6, FR-09.7
-         */
-        ngx_log_error(NGX_LOG_CRIT, cycle->log, 0,
-                      "markdown filter: failed to initialize converter in worker process, "
-                      "category=system");
-        return NGX_ERROR;
-    }
-    
-    /*
-     * Log successful initialization at info level.
-     * 
-     * This helps administrators verify that the module is properly
-     * initialized in each worker process. The log message includes
-     * the process ID to distinguish between workers.
-     */
-    ngx_log_error(NGX_LOG_INFO, cycle->log, 0,
-                  "markdown filter: converter initialized in worker process (pid: %P)",
-                  ngx_pid);
-    
-    /*
-     * Log decompression support information.
-     * 
-     * This informs operators about which compression formats are supported
-     * at startup. The brotli status depends on whether NGX_HTTP_BROTLI
-     * was defined at compile time.
-     * 
-     * Requirements: 12.1, 12.2, 12.3, 12.4, 12.5
-     */
-#ifdef NGX_HTTP_BROTLI
-    ngx_log_error(NGX_LOG_INFO, cycle->log, 0,
-                  "markdown filter: decompression support: gzip=yes, deflate=yes, brotli=yes");
-#else
-    ngx_log_error(NGX_LOG_INFO, cycle->log, 0,
-                  "markdown filter: decompression support: gzip=yes, deflate=yes, brotli=no");
-#endif
-    
-    return NGX_OK;
-}
-
-/*
- * Worker process cleanup
- *
- * Called once per worker process when NGINX shuts down or reloads.
- * Cleans up the Rust converter instance for this worker.
- *
- * This function is called during graceful shutdown, after all active
- * requests have completed. It ensures proper resource cleanup to prevent
- * memory leaks.
- *
- * @param cycle  The NGINX cycle structure
- */
-static void
-ngx_http_markdown_exit_worker(ngx_cycle_t *cycle)
-{
-    /*
-     * Check if converter was initialized.
-     * 
-     * The converter might be NULL if:
-     * - Worker initialization failed
-     * - This function is called multiple times (shouldn't happen)
-     * - Module was never used in this worker
-     */
-    if (ngx_http_markdown_converter == NULL) {
-        /*
-         * No converter to clean up.
-         * 
-         * This is not an error - it just means the converter was never
-         * initialized (perhaps due to initialization failure or module
-         * not being used). We log at debug level for troubleshooting.
-         */
-        ngx_log_error(NGX_LOG_DEBUG, cycle->log, 0,
-                      "markdown filter: no converter to clean up in worker process");
-        return;
-    }
-    
-    /*
-     * Free the converter instance.
-     * 
-     * This calls the Rust FFI function to properly deallocate the
-     * converter and all its internal resources. The function is safe
-     * to call even if there are no active conversions (which there
-     * shouldn't be at this point, since we're shutting down).
-     */
-    markdown_converter_free(ngx_http_markdown_converter);
-    
-    /*
-     * Set the global pointer to NULL.
-     * 
-     * This prevents use-after-free bugs if this function is somehow
-     * called multiple times, or if there are any lingering references
-     * to the converter during shutdown.
-     */
-    ngx_http_markdown_converter = NULL;
-    
-    /*
-     * Log successful cleanup at info level.
-     * 
-     * This helps administrators verify that resources are being properly
-     * cleaned up during shutdown or reload. The log message includes
-     * the process ID to distinguish between workers.
-     */
-    ngx_log_error(NGX_LOG_INFO, cycle->log, 0,
-                  "markdown filter: converter cleaned up in worker process (pid: %P)",
-                  ngx_pid);
-}
-
-/*
- * Header filter
- *
- * Called when response headers are being sent.
- * Checks if conversion should be attempted based on:
- * - Accept header (text/markdown)
- * - Configuration (enabled/disabled)
- * - Request method (GET/HEAD)
- * - Response status and content type
- * 
- * If conversion is eligible, sets up conversion context.
- * Otherwise, passes through to next filter.
- *
- * This is part of Task 14.4 setup - creates context for body buffering.
- */
-static ngx_int_t
-ngx_http_markdown_header_filter(ngx_http_request_t *r)
-{
-    ngx_http_markdown_ctx_t         *ctx;
-    ngx_http_markdown_conf_t        *conf;
-    ngx_http_markdown_eligibility_t  eligibility;
-    ngx_flag_t                       filter_enabled;
-    ngx_int_t                        should_convert;
-
-    /* Get module configuration */
-    conf = ngx_http_get_module_loc_conf(r, ngx_http_markdown_filter_module);
-    if (conf == NULL) {
-        /* Module not configured, pass through */
-        return ngx_http_next_header_filter(r);
-    }
-
-    /*
-     * Resolve markdown_filter once in header phase and cache the result in
-     * request context. Body phase must reuse this decision to avoid
-     * header/body inconsistencies for dynamic variables.
-     */
-    filter_enabled = ngx_http_markdown_is_enabled(r, conf);
-    if (!filter_enabled) {
-        /* Module disabled, pass through */
-        return ngx_http_next_header_filter(r);
-    }
-
-    /* Check if client wants Markdown (Accept header) */
-    should_convert = ngx_http_markdown_should_convert(r, conf);
-    if (!should_convert) {
-        /* Client doesn't want Markdown, pass through */
-        return ngx_http_next_header_filter(r);
-    }
-
-    /* Check if response is eligible for conversion */
-    eligibility = ngx_http_markdown_check_eligibility(r, conf, filter_enabled);
-    if (eligibility != NGX_HTTP_MARKDOWN_ELIGIBLE) {
-        /* Not eligible, pass through */
-        ngx_log_debug1(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                      "markdown filter: response not eligible: %V",
-                      ngx_http_markdown_eligibility_string(eligibility));
-        return ngx_http_next_header_filter(r);
-    }
-
-    /* Create request context for buffering */
-    ctx = ngx_pcalloc(r->pool, sizeof(ngx_http_markdown_ctx_t));
-    if (ctx == NULL) {
-        /*
-         * Context allocation failed - critical system error.
-         * This means we're out of memory.
-         * 
-         * Log level: NGX_LOG_CRIT (critical system failure)
-         * Requirements: FR-09.5, FR-09.6, FR-09.7
-         */
-        ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                     "markdown filter: failed to allocate context, category=system");
-        return ngx_http_next_header_filter(r);
-    }
-
-    /* Initialize context */
-    ctx->request = r;
-    ctx->filter_enabled = filter_enabled;
-    ctx->eligible = 1;
-    ctx->buffer_initialized = 0;
-    ctx->headers_forwarded = 0;
-    ctx->conversion_attempted = 0;
-    ctx->conversion_succeeded = 0;
-    
-    /* Initialize decompression state (Task 2.4 - Fast path initialization)
-     * 
-     * For uncompressed content, decompression_needed remains 0, ensuring
-     * zero overhead in the body filter (no decompression functions called).
-     * 
-     * Requirements: 4.2, 10.3
-     */
-    ctx->compression_type = NGX_HTTP_MARKDOWN_COMPRESSION_NONE;
-    ctx->decompression_needed = 0;
-    ctx->decompression_done = 0;
-    ctx->compressed_size = 0;
-    ctx->decompressed_size = 0;
-
-    /* Set context for this request */
-    ngx_http_set_ctx(r, ctx, ngx_http_markdown_filter_module);
-
-    /*
-     * Detect compression type if auto_decompress is enabled (Task 2.1, 4.2)
-     * 
-     * Fast path: If compression_type == NONE, decompression_needed stays 0
-     * Slow path: If compression detected, decompression_needed is set to 1
-     * Special case: If compression_type == UNKNOWN, trigger fail-open immediately
-     * 
-     * Requirements: 1.1, 1.6, 4.2, 8.1, 10.3, 11.1, 11.5
-     */
-    if (conf->auto_decompress) {
-        ctx->compression_type = ngx_http_markdown_detect_compression(r);
-        
-        if (ctx->compression_type == NGX_HTTP_MARKDOWN_COMPRESSION_UNKNOWN) {
-            /*
-             * Unsupported compression format detected (Task 4.2)
-             * 
-             * This is an expected degradation scenario, not a failure.
-             * We gracefully degrade by returning the original content.
-             * 
-             * Note: The warning log has already been emitted by
-             * ngx_http_markdown_detect_compression() with the format name.
-             * 
-             * Requirements: 1.6, 11.5
-             */
-            ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                         "markdown filter: unsupported compression format detected, "
-                         "returning original content (fail-open)");
-            
-            /* Set eligible = 0 to trigger fail-open (return original content) */
-            ctx->eligible = 0;
-            
-            /* Don't set decompression_needed - we're not attempting decompression */
-            /* Don't increment failure counter - this is expected degradation */
-            
-        } else if (ctx->compression_type != NGX_HTTP_MARKDOWN_COMPRESSION_NONE) {
-            /* Supported compression format - set flag for decompression */
-            ctx->decompression_needed = 1;
-            
-            ngx_log_debug1(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                          "markdown filter: decompression detected compression type: %d",
-                          ctx->compression_type);
-        }
-    }
-
-    /*
-     * Request in-memory buffers from upstream filters/modules.
-     *
-     * Static file responses may otherwise arrive as file-backed buffers
-     * (sendfile path), where `buf->pos..last` is empty and the payload is
-     * only described by file offsets. This filter buffers and converts the
-     * response body in userspace, so it requires in-memory data.
-     */
-    r->filter_need_in_memory = 1;
-
-    ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: response eligible for conversion, "
-                  "context initialized");
-
-    /*
-     * Defer downstream header emission until the body filter completes
-     * conversion (or decides to fail-open). This allows the module to set
-     * accurate Content-Type / Content-Length / ETag headers based on the
-     * converted Markdown output.
-     */
-    return NGX_OK;
-}
-
-/*
- * Construct base URL for resolving relative URLs
- *
- * This function constructs the base URL using the following priority order:
- * 1. X-Forwarded-Proto + X-Forwarded-Host (reverse proxy scenario)
- * 2. r->schema + r->headers_in.server (direct connection)
- * 3. server_name from configuration (fallback)
- *
- * The base_url is used by the Rust conversion engine to resolve relative URLs
- * in HTML to absolute URLs in the Markdown output.
- *
- * Format: scheme://host/uri
- * Example: https://example.com/docs/page.html
- *
- * @param r     The request structure
- * @param pool  Memory pool for allocation
- * @param base_url  Output parameter for constructed base URL
- * @return      NGX_OK on success, NGX_ERROR on failure
- *
- * Requirements: Design - URL Resolution, NGINX Integration
- * Task: 14.8 Implement base_url construction with X-Forwarded headers priority
- */
-static u_char ngx_http_markdown_hdr_x_forwarded_proto[] = "X-Forwarded-Proto";
-static u_char ngx_http_markdown_hdr_x_forwarded_host[] = "X-Forwarded-Host";
-static u_char ngx_http_markdown_scheme_http[] = "http";
-static u_char ngx_http_markdown_scheme_https[] = "https";
-
-static ngx_str_t *
-ngx_http_markdown_find_request_header_value(ngx_http_request_t *r,
-                                            u_char *name,
-                                            size_t name_len)
-{
-    ngx_list_part_t  *part;
-
-    if (r->headers_in.headers.part.nelts == 0) {
-        return NULL;
-    }
-
-    for (part = &r->headers_in.headers.part; part != NULL; part = part->next) {
-        ngx_table_elt_t  *headers;
-        ngx_uint_t        i;
-
-        headers = part->elts;
-        for (i = 0; i < part->nelts; i++) {
-            if (headers[i].key.len == name_len
-                && ngx_strncasecmp(headers[i].key.data, name, name_len) == 0)
-            {
-                return &headers[i].value;
-            }
-        }
-    }
-
-    return NULL;
-}
-
-static ngx_flag_t
-ngx_http_markdown_scheme_is_http_family(const ngx_str_t *scheme)
-{
-    return (scheme->len == sizeof(ngx_http_markdown_scheme_http) - 1
-            && ngx_strncasecmp(scheme->data,
-                               ngx_http_markdown_scheme_http,
-                               sizeof(ngx_http_markdown_scheme_http) - 1) == 0)
-        || (scheme->len == sizeof(ngx_http_markdown_scheme_https) - 1
-            && ngx_strncasecmp(scheme->data,
-                               ngx_http_markdown_scheme_https,
-                               sizeof(ngx_http_markdown_scheme_https) - 1) == 0);
-}
-
-static ngx_int_t
-ngx_http_markdown_select_base_url_parts(ngx_http_request_t *r,
-                                        ngx_str_t *scheme,
-                                        ngx_str_t *host)
-{
-    ngx_str_t                 *x_forwarded_proto;
-    ngx_str_t                 *x_forwarded_host;
-    ngx_http_core_srv_conf_t *cscf;
-
-    x_forwarded_proto = ngx_http_markdown_find_request_header_value(
-        r,
-        ngx_http_markdown_hdr_x_forwarded_proto,
-        sizeof(ngx_http_markdown_hdr_x_forwarded_proto) - 1);
-    x_forwarded_host = ngx_http_markdown_find_request_header_value(
-        r,
-        ngx_http_markdown_hdr_x_forwarded_host,
-        sizeof(ngx_http_markdown_hdr_x_forwarded_host) - 1);
-
-    if (x_forwarded_proto != NULL
-        && x_forwarded_host != NULL
-        && x_forwarded_proto->len > 0
-        && x_forwarded_host->len > 0
-        && ngx_http_markdown_scheme_is_http_family(x_forwarded_proto))
-    {
-        *scheme = *x_forwarded_proto;
-        *host = *x_forwarded_host;
-        return NGX_OK;
-    }
-
-    if (r->schema.len > 0 && r->headers_in.server.len > 0) {
-        *scheme = r->schema;
-        *host = r->headers_in.server;
-        return NGX_OK;
-    }
-
-    cscf = ngx_http_get_module_srv_conf(r, ngx_http_core_module);
-    if (cscf != NULL && cscf->server_name.len > 0) {
-        if (r->schema.len > 0) {
-            *scheme = r->schema;
-        } else {
-            scheme->data = ngx_http_markdown_scheme_http;
-            scheme->len = sizeof(ngx_http_markdown_scheme_http) - 1;
-        }
-        *host = cscf->server_name;
-        return NGX_OK;
-    }
-
-    ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                 "markdown filter: unable to construct base_url, "
-                 "no valid scheme/host available");
-    return NGX_ERROR;
-}
-
-static ngx_int_t
-ngx_http_markdown_base_url_add_len(ngx_http_request_t *r,
-                                   size_t *total,
-                                   size_t delta,
-                                   const char *segment)
-{
-    if (*total > ((size_t) -1) - delta) {
-        ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                     "markdown filter: base_url length overflow (%s)",
-                     segment);
-        return NGX_ERROR;
-    }
-    *total += delta;
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_construct_base_url(ngx_http_request_t *r, ngx_pool_t *pool,
-    ngx_str_t *base_url)
-{
-    ngx_str_t  scheme;
-    ngx_str_t  host;
-    u_char    *p;
-    size_t     len;
-
-    /* Initialize output */
-    base_url->data = NULL;
-    base_url->len = 0;
-
-    if (ngx_http_markdown_select_base_url_parts(r, &scheme, &host) != NGX_OK) {
-        return NGX_ERROR;
-    }
-
-    len = 0;
-    if (ngx_http_markdown_base_url_add_len(r, &len, scheme.len, "scheme") != NGX_OK) {
-        return NGX_ERROR;
-    }
-    if (ngx_http_markdown_base_url_add_len(r, &len, sizeof("://") - 1, "delimiter") != NGX_OK) {
-        return NGX_ERROR;
-    }
-    if (ngx_http_markdown_base_url_add_len(r, &len, host.len, "host") != NGX_OK) {
-        return NGX_ERROR;
-    }
-    if (ngx_http_markdown_base_url_add_len(r, &len, r->uri.len, "uri") != NGX_OK) {
-        return NGX_ERROR;
-    }
-    
-    p = ngx_pnalloc(pool, len);
-    if (p == NULL) {
-        /*
-         * Memory allocation failed for base_url - critical system error.
-         * 
-         * Log level: NGX_LOG_CRIT (critical system failure)
-         * Requirements: FR-09.5, FR-09.6, FR-09.7
-         */
-        ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                     "markdown filter: failed to allocate memory for base_url, category=system");
-        return NGX_ERROR;
-    }
-
-    base_url->data = p;
-    
-    /* Copy scheme */
-    p = ngx_cpymem(p, scheme.data, scheme.len);
-    
-    /* Add :// */
-    *p++ = ':';
-    *p++ = '/';
-    *p++ = '/';
-    
-    /* Copy host */
-    p = ngx_cpymem(p, host.data, host.len);
-    
-    /* Copy URI */
-    p = ngx_cpymem(p, r->uri.data, r->uri.len);
-    
-    base_url->len = p - base_url->data;
-
-    ngx_log_debug1(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: constructed base_url: \"%V\"", base_url);
-
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_fail_open_buffered_response(ngx_http_request_t *r,
-                                              ngx_http_markdown_ctx_t *ctx,
-                                              const char *debug_message)
-{
-    if (debug_message != NULL) {
-        ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0, debug_message);
-    }
-
-    ctx->eligible = 0;
-    r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-
-    if (ngx_http_markdown_forward_headers(r, ctx) != NGX_OK) {
-        return NGX_ERROR;
-    }
-
-    return ngx_http_markdown_send_buffered_original_response(r, ctx);
-}
-
-static ngx_int_t
-ngx_http_markdown_reject_or_fail_open_buffered_response(
-    ngx_http_request_t *r,
-    ngx_http_markdown_ctx_t *ctx,
-    ngx_http_markdown_conf_t *conf,
-    const char *debug_message)
-{
-    if (conf->on_error == NGX_HTTP_MARKDOWN_ON_ERROR_REJECT) {
-        return NGX_ERROR;
-    }
-
-    return ngx_http_markdown_fail_open_buffered_response(r, ctx, debug_message);
-}
-
-static ngx_int_t
-ngx_http_markdown_prepare_compressed_chain(ngx_http_request_t *r,
-                                           ngx_http_markdown_ctx_t *ctx,
-                                           ngx_http_markdown_conf_t *conf,
-                                           ngx_chain_t **compressed_chain)
-{
-    ngx_buf_t *compressed_buf;
-
-    compressed_buf = ngx_create_temp_buf(r->pool, ctx->buffer.size);
-    if (compressed_buf == NULL) {
-        ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                     "markdown filter: failed to create compressed buffer, "
-                     "category=system");
-        return ngx_http_markdown_reject_or_fail_open_buffered_response(
-            r, ctx, conf,
-            "markdown filter: fail-open strategy - returning original content");
-    }
-
-    if (ctx->buffer.size > 0) {
-        if (ctx->buffer.data == NULL) {
-            ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                         "markdown filter: buffered payload pointer is NULL with non-zero size, "
-                         "size=%uz, category=system",
-                         ctx->buffer.size);
-            return ngx_http_markdown_reject_or_fail_open_buffered_response(
-                r, ctx, conf, NULL);
-        }
-
-        ngx_memcpy(compressed_buf->pos, ctx->buffer.data, ctx->buffer.size);
-    }
-    compressed_buf->last = compressed_buf->pos + ctx->buffer.size;
-    compressed_buf->last_buf = 1;
-
-    *compressed_chain = ngx_alloc_chain_link(r->pool);
-    if (*compressed_chain == NULL) {
-        ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                     "markdown filter: failed to allocate chain link, "
-                     "category=system");
-        return ngx_http_markdown_reject_or_fail_open_buffered_response(
-            r, ctx, conf, NULL);
-    }
-
-    (*compressed_chain)->buf = compressed_buf;
-    (*compressed_chain)->next = NULL;
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_apply_decompressed_payload(ngx_http_request_t *r,
-                                             ngx_http_markdown_ctx_t *ctx,
-                                             ngx_http_markdown_conf_t *conf,
-                                             ngx_chain_t *decompressed_chain)
-{
-    const u_char *decompressed_data;
-    u_char       *target_data;
-
-    if (decompressed_chain == NULL || decompressed_chain->buf == NULL) {
-        const ngx_str_t *compression_name;
-
-        compression_name = ngx_http_markdown_compression_name(
-            ctx->compression_type);
-
-        ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                     "markdown filter: decompression returned NULL chain, "
-                     "compression=%V, category=system",
-                     compression_name);
-
-        ngx_atomic_fetch_add(&ngx_http_markdown_metrics.decompressions_failed, 1);
-        return ngx_http_markdown_reject_or_fail_open_buffered_response(
-            r, ctx, conf, NULL);
-    }
-
-    ctx->decompressed_size = decompressed_chain->buf->last - decompressed_chain->buf->pos;
-    decompressed_data = decompressed_chain->buf->pos;
-    target_data = ctx->buffer.data;
-
-    if (ctx->decompressed_size > ctx->buffer.capacity) {
-        u_char *new_data = ngx_alloc(ctx->decompressed_size, r->connection->log);
-        if (new_data == NULL) {
-            const ngx_str_t *compression_name;
-
-            compression_name = ngx_http_markdown_compression_name(
-                ctx->compression_type);
-
-            ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                         "markdown filter: failed to allocate decompressed buffer, "
-                         "compression=%V, size=%uz, category=system",
-                         compression_name, ctx->decompressed_size);
-
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.decompressions_failed, 1);
-            return ngx_http_markdown_reject_or_fail_open_buffered_response(
-                r, ctx, conf, NULL);
-        }
-
-        target_data = new_data;
-    }
-
-    if (ctx->decompressed_size > 0 && target_data != decompressed_data) {
-        ngx_memcpy(target_data, decompressed_data, ctx->decompressed_size);
-    }
-
-    if (target_data != ctx->buffer.data) {
-        if (ctx->buffer.data != NULL) {
-            ngx_free(ctx->buffer.data);
-        }
-
-        ctx->buffer.data = target_data;
-        ctx->buffer.capacity = ctx->decompressed_size;
-    }
-
-    ctx->buffer.size = ctx->decompressed_size;
-    ctx->decompression_done = 1;
-    return NGX_OK;
-}
-
-static void
-ngx_http_markdown_record_decompression_success(ngx_http_request_t *r,
-                                               ngx_http_markdown_ctx_t *ctx)
-{
-    ngx_atomic_fetch_add(&ngx_http_markdown_metrics.decompressions_succeeded, 1);
-    switch (ctx->compression_type) {
-        case NGX_HTTP_MARKDOWN_COMPRESSION_GZIP:
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.decompressions_gzip, 1);
-            break;
-        case NGX_HTTP_MARKDOWN_COMPRESSION_DEFLATE:
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.decompressions_deflate, 1);
-            break;
-        case NGX_HTTP_MARKDOWN_COMPRESSION_BROTLI:
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.decompressions_brotli, 1);
-            break;
-        default:
-            break;
-    }
-
-    ngx_log_error(NGX_LOG_INFO, r->connection->log, 0,
-                 "markdown filter: decompression succeeded, "
-                 "compression=%d, compressed=%uz bytes, decompressed=%uz bytes, ratio=%.1fx",
-                 ctx->compression_type, ctx->compressed_size, ctx->decompressed_size,
-                 (float) ctx->decompressed_size / ctx->compressed_size);
-
-    ngx_http_markdown_remove_content_encoding(r);
-    ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: removed Content-Encoding header after decompression");
-}
-
-static ngx_int_t
-ngx_http_markdown_resolve_conditional_result(ngx_http_request_t *r,
-                                             ngx_http_markdown_ctx_t *ctx,
-                                             ngx_http_markdown_conf_t *conf,
-                                             struct MarkdownResult *result,
-                                             ngx_msec_t *elapsed_ms,
-                                             ngx_flag_t *has_result)
-{
-    struct MarkdownResult *conditional_result;
-    ngx_int_t              rc;
-
-    conditional_result = NULL;
-    *has_result = 0;
-
-    rc = ngx_http_markdown_handle_if_none_match(
-        r, conf, ctx, ngx_http_markdown_converter, &conditional_result);
-
-    if (rc == NGX_HTTP_NOT_MODIFIED) {
-        ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                      "markdown filter: If-None-Match matched, sending 304 Not Modified");
-
-        rc = ngx_http_markdown_send_304(r, conditional_result);
-        if (conditional_result != NULL) {
-            markdown_result_free(conditional_result);
-        }
-
-        r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-        return NGX_DONE;
-    }
-
-    if (rc == NGX_ERROR) {
-        ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                     "markdown filter: error during If-None-Match processing");
-
-        if (conditional_result != NULL) {
-            markdown_result_free(conditional_result);
-        }
-
-        return ngx_http_markdown_reject_or_fail_open_buffered_response(
-            r, ctx, conf,
-            "markdown filter: fail-open strategy - returning original HTML");
-    }
-
-    if (rc == NGX_DECLINED && conditional_result != NULL) {
-        ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                      "markdown filter: If-None-Match did not match, using existing conversion");
-
-        ngx_memcpy(result, conditional_result, sizeof(struct MarkdownResult));
-        ngx_pfree(r->pool, conditional_result);
-
-        *elapsed_ms = 0;
-        *has_result = 1;
-    }
-
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_prepare_conversion_options(ngx_http_request_t *r,
-                                             ngx_http_markdown_conf_t *conf,
-                                             struct MarkdownOptions *options)
-{
-    ngx_str_t base_url;
-
-    ngx_memzero(options, sizeof(struct MarkdownOptions));
-    options->flavor = conf->flavor;
-    options->timeout_ms = conf->timeout;
-    options->generate_etag = conf->generate_etag;
-    options->estimate_tokens = conf->token_estimate;
-    options->front_matter = conf->front_matter;
-    options->content_type = NULL;
-    options->content_type_len = 0;
-    options->base_url = NULL;
-    options->base_url_len = 0;
-
-    if (r->headers_out.content_type.len > 0) {
-        options->content_type = r->headers_out.content_type.data;
-        options->content_type_len = r->headers_out.content_type.len;
-    }
-
-    if (ngx_http_markdown_construct_base_url(r, r->pool, &base_url) == NGX_OK) {
-        options->base_url = base_url.data;
-        options->base_url_len = base_url.len;
-    } else {
-        ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                      "markdown filter: continuing conversion without base_url");
-    }
-
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_handle_conversion_failure(ngx_http_request_t *r,
-                                            ngx_http_markdown_ctx_t *ctx,
-                                            ngx_http_markdown_conf_t *conf,
-                                            struct MarkdownResult *result,
-                                            ngx_msec_t elapsed_ms)
-{
-    ngx_http_markdown_error_category_t error_category;
-    const ngx_str_t                  *category_str;
-
-    error_category = ngx_http_markdown_classify_error(result->error_code);
-    category_str = ngx_http_markdown_error_category_string(error_category);
-    ngx_atomic_fetch_add(&ngx_http_markdown_metrics.conversions_failed, 1);
-
-    switch (error_category) {
-        case NGX_HTTP_MARKDOWN_ERROR_CONVERSION:
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.failures_conversion, 1);
-            break;
-        case NGX_HTTP_MARKDOWN_ERROR_RESOURCE_LIMIT:
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.failures_resource_limit, 1);
-            break;
-        case NGX_HTTP_MARKDOWN_ERROR_SYSTEM:
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.failures_system, 1);
-            break;
-    }
-
-    ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                 "markdown filter: conversion failed, "
-                 "error_code=%ud, category=%V, message=\"%*s\", elapsed_ms=%M",
-                 result->error_code,
-                 category_str,
-                 (result->error_message != NULL) ? (ngx_int_t) result->error_len : 0,
-                 (result->error_message != NULL) ? result->error_message : ngx_http_markdown_empty_string,
-                 elapsed_ms);
-
-    markdown_result_free(result);
-    return ngx_http_markdown_reject_or_fail_open_buffered_response(
-        r, ctx, conf,
-        "markdown filter: fail-open strategy - returning original HTML");
-}
-
-static ngx_int_t
-ngx_http_markdown_validate_conversion_result(ngx_http_request_t *r,
-                                             ngx_http_markdown_ctx_t *ctx,
-                                             ngx_http_markdown_conf_t *conf,
-                                             struct MarkdownResult *result)
-{
-    if ((result->markdown == NULL && result->markdown_len > 0)
-        || (result->error_message == NULL && result->error_len > 0)
-        || (result->etag == NULL && result->etag_len > 0))
-    {
-        ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                     "markdown filter: invalid FFI result invariants: "
-                     "markdown=%p markdown_len=%uz etag=%p etag_len=%uz "
-                     "error_message=%p error_len=%uz",
-                     result->markdown, result->markdown_len,
-                     result->etag, result->etag_len,
-                     result->error_message, result->error_len);
-        markdown_result_free(result);
-        return ngx_http_markdown_reject_or_fail_open_buffered_response(
-            r, ctx, conf, NULL);
-    }
-
-    return NGX_OK;
-}
-
-static void
-ngx_http_markdown_record_conversion_success(ngx_http_markdown_ctx_t *ctx,
-                                            struct MarkdownResult *result,
-                                            ngx_msec_t elapsed_ms)
-{
-    ctx->conversion_succeeded = 1;
-    ngx_atomic_fetch_add(&ngx_http_markdown_metrics.conversions_succeeded, 1);
-    ngx_atomic_fetch_add(&ngx_http_markdown_metrics.input_bytes, ctx->buffer.size);
-    ngx_atomic_fetch_add(&ngx_http_markdown_metrics.output_bytes, result->markdown_len);
-    ngx_atomic_fetch_add(&ngx_http_markdown_metrics.conversion_time_sum_ms, elapsed_ms);
-}
-
-static ngx_int_t
-ngx_http_markdown_execute_conversion(ngx_http_request_t *r,
-                                     ngx_http_markdown_ctx_t *ctx,
-                                     ngx_http_markdown_conf_t *conf,
-                                     struct MarkdownResult *result,
-                                     ngx_msec_t *elapsed_ms)
-{
-    struct MarkdownOptions  options;
-    ngx_time_t             *tp;
-    ngx_msec_t              start_time;
-    ngx_msec_t              end_time;
-    ngx_int_t               rc;
-
-    if (ngx_http_markdown_converter == NULL) {
-        ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                     "markdown filter: converter not initialized, category=system");
-        return ngx_http_markdown_reject_or_fail_open_buffered_response(
-            r, ctx, conf,
-            "markdown filter: fail-open strategy - returning original HTML");
-    }
-
-    ngx_http_markdown_prepare_conversion_options(r, conf, &options);
-
-    tp = ngx_timeofday();
-    start_time = (ngx_msec_t) (tp->sec * 1000 + tp->msec);
-
-    ngx_memzero(result, sizeof(struct MarkdownResult));
-    markdown_convert(ngx_http_markdown_converter,
-                    ctx->buffer.data,
-                    ctx->buffer.size,
-                    &options,
-                    result);
-
-    tp = ngx_timeofday();
-    end_time = (ngx_msec_t) (tp->sec * 1000 + tp->msec);
-    *elapsed_ms = end_time - start_time;
-
-    if (result->error_code != ERROR_SUCCESS) {
-        return ngx_http_markdown_handle_conversion_failure(
-            r, ctx, conf, result, *elapsed_ms);
-    }
-
-    rc = ngx_http_markdown_validate_conversion_result(r, ctx, conf, result);
-    if (rc != NGX_OK) {
-        return rc;
-    }
-
-    ngx_http_markdown_record_conversion_success(ctx, result, *elapsed_ms);
-    return NGX_OK;
-}
-
-static void
-ngx_http_markdown_prepare_head_output_buffer(ngx_http_request_t *r,
-                                             ngx_buf_t *b,
-                                             struct MarkdownResult *result)
-{
-    ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: HEAD request - omitting response body");
-
-    b->pos = NULL;
-    b->last = NULL;
-    b->memory = 0;
-    b->last_buf = (r == r->main) ? 1 : 0;
-    b->last_in_chain = 1;
-    markdown_result_free(result);
-}
-
-static ngx_int_t
-ngx_http_markdown_prepare_body_output_buffer(ngx_http_request_t *r,
-                                             ngx_buf_t *b,
-                                             struct MarkdownResult *result)
-{
-    if (result->markdown_len > 0) {
-        b->pos = ngx_pnalloc(r->pool, result->markdown_len);
-        if (b->pos == NULL) {
-            ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                         "markdown filter: failed to allocate output memory, category=system");
-            markdown_result_free(result);
-            return NGX_ERROR;
-        }
-
-        ngx_memcpy(b->pos, result->markdown, result->markdown_len);
-        b->last = b->pos + result->markdown_len;
-        b->memory = 1;
-    } else {
-        b->pos = NULL;
-        b->last = NULL;
-        b->memory = 0;
-    }
-
-    b->last_buf = (r == r->main) ? 1 : 0;
-    b->last_in_chain = 1;
-    markdown_result_free(result);
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_send_conversion_output(ngx_http_request_t *r,
-                                         ngx_http_markdown_ctx_t *ctx,
-                                         ngx_http_markdown_conf_t *conf,
-                                         struct MarkdownResult *result,
-                                         ngx_msec_t elapsed_ms)
-{
-    ngx_int_t   rc;
-    ngx_chain_t *out;
-    ngx_buf_t   *b;
-
-    ngx_log_debug3(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: conversion succeeded, "
-                  "input: %uz bytes, output: %uz bytes, elapsed: %M ms",
-                  ctx->buffer.size, result->markdown_len, elapsed_ms);
-
-    rc = ngx_http_markdown_update_headers(r, result, conf);
-    if (rc != NGX_OK) {
-        ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                     "markdown filter: failed to update response headers, category=system");
-        markdown_result_free(result);
-        return NGX_ERROR;
-    }
-
-    rc = ngx_http_markdown_forward_headers(r, ctx);
-    if (rc != NGX_OK) {
-        markdown_result_free(result);
-        return rc;
-    }
-
-    b = ngx_pcalloc(r->pool, sizeof(ngx_buf_t));
-    if (b == NULL) {
-        ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                     "markdown filter: failed to allocate output buffer, category=system");
-        markdown_result_free(result);
-        return NGX_ERROR;
-    }
-
-    if (r->method == NGX_HTTP_HEAD) {
-        ngx_http_markdown_prepare_head_output_buffer(r, b, result);
-    } else {
-        rc = ngx_http_markdown_prepare_body_output_buffer(r, b, result);
-        if (rc != NGX_OK) {
-            return rc;
-        }
-    }
-
-    out = ngx_alloc_chain_link(r->pool);
-    if (out == NULL) {
-        ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                     "markdown filter: failed to allocate output chain, category=system");
-        return NGX_ERROR;
-    }
-    out->buf = b;
-    out->next = NULL;
-
-    return ngx_http_next_body_filter(r, out);
-}
-
-static ngx_int_t
-ngx_http_markdown_handle_buffer_init_failure(ngx_http_request_t *r,
-                                             ngx_http_markdown_ctx_t *ctx,
-                                             ngx_http_markdown_conf_t *conf,
-                                             ngx_chain_t *in)
-{
-    ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                 "markdown filter: failed to initialize buffer, category=system");
-
-    if (conf->on_error == NGX_HTTP_MARKDOWN_ON_ERROR_REJECT) {
-        return NGX_ERROR;
-    }
-
-    ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: fail-open strategy - returning original HTML");
-    ctx->eligible = 0;
-    r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-    if (ngx_http_markdown_forward_headers(r, ctx) != NGX_OK) {
-        return NGX_ERROR;
-    }
-
-    return ngx_http_next_body_filter(r, in);
-}
-
-static ngx_int_t
-ngx_http_markdown_handle_buffer_append_failure(ngx_http_request_t *r,
-                                               ngx_http_markdown_ctx_t *ctx,
-                                               ngx_http_markdown_conf_t *conf,
-                                               ngx_chain_t *cl,
-                                               size_t chunk_size)
-{
-    size_t attempted_size;
-
-    attempted_size = ctx->buffer.max_size;
-    if (ctx->buffer.size <= ctx->buffer.max_size
-        && chunk_size <= (ctx->buffer.max_size - ctx->buffer.size))
-    {
-        attempted_size = ctx->buffer.size + chunk_size;
-    }
-
-    ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                 "markdown filter: response size exceeds limit, "
-                 "size=%uz bytes, max=%uz bytes, category=resource_limit",
-                 attempted_size, conf->max_size);
-
-    if (conf->on_error == NGX_HTTP_MARKDOWN_ON_ERROR_REJECT) {
-        return NGX_ERROR;
-    }
-
-    ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: fail-open strategy - returning original HTML");
-    ctx->eligible = 0;
-    r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-    if (ngx_http_markdown_forward_headers(r, ctx) != NGX_OK) {
-        return NGX_ERROR;
-    }
-
-    return ngx_http_markdown_fail_open_with_buffered_prefix(r, ctx, cl);
-}
-
-static ngx_int_t
-ngx_http_markdown_append_buffered_chunk(ngx_http_request_t *r,
-                                        ngx_http_markdown_ctx_t *ctx,
-                                        ngx_http_markdown_conf_t *conf,
-                                        ngx_chain_t *cl)
-{
-    ngx_int_t rc;
-    size_t    chunk_size;
-
-    if (cl->buf == NULL) {
-        return NGX_OK;
-    }
-
-    chunk_size = cl->buf->last - cl->buf->pos;
-    if (chunk_size == 0) {
-        return NGX_OK;
-    }
-
-    rc = ngx_http_markdown_buffer_append(&ctx->buffer, cl->buf->pos, chunk_size);
-    if (rc != NGX_OK) {
-        return ngx_http_markdown_handle_buffer_append_failure(
-            r, ctx, conf, cl, chunk_size);
-    }
-
-    cl->buf->pos = cl->buf->last;
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_body_filter_buffer_input(ngx_http_request_t *r,
-                                           ngx_chain_t *in,
-                                           ngx_http_markdown_ctx_t *ctx,
-                                           ngx_http_markdown_conf_t *conf)
-{
-    ngx_chain_t  *cl;
-    ngx_int_t     rc;
-    ngx_flag_t    last_buf;
-    size_t        reserve_hint;
-    off_t         content_length_n;
-
-    if (!ctx->buffer_initialized) {
-        rc = ngx_http_markdown_buffer_init(&ctx->buffer, conf->max_size, r->pool);
-        if (rc != NGX_OK) {
-            return ngx_http_markdown_handle_buffer_init_failure(r, ctx, conf, in);
-        }
-        ctx->buffer_initialized = 1;
-
-        content_length_n = r->headers_out.content_length_n;
-        if (content_length_n > 0 && content_length_n <= (off_t) conf->max_size) {
-            reserve_hint = (size_t) content_length_n;
-            if (reserve_hint > NGX_HTTP_MARKDOWN_PRERESERVE_LIMIT) {
-                reserve_hint = NGX_HTTP_MARKDOWN_PRERESERVE_LIMIT;
-            }
-
-            if (ngx_http_markdown_buffer_reserve(&ctx->buffer, reserve_hint) != NGX_OK) {
-                ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                             "markdown filter: failed to pre-reserve %uz bytes buffer capacity",
-                             reserve_hint);
-            }
-        }
-    }
-
-    last_buf = 0;
-    for (cl = in; cl != NULL; cl = cl->next) {
-        rc = ngx_http_markdown_append_buffered_chunk(r, ctx, conf, cl);
-        if (rc != NGX_OK) {
-            return rc;
-        }
-
-        if (cl->buf != NULL
-            && (cl->buf->last_buf || (r != r->main && cl->buf->last_in_chain)))
-        {
-            last_buf = 1;
-            break;
-        }
-    }
-
-    if (!last_buf) {
-        r->buffered |= NGX_HTTP_MARKDOWN_BUFFERED;
-        return NGX_AGAIN;
-    }
-
-    r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_body_filter_decompress_if_needed(ngx_http_request_t *r,
-                                                   ngx_http_markdown_ctx_t *ctx,
-                                                   ngx_http_markdown_conf_t *conf)
-{
-    if (ctx->decompression_needed && !ctx->decompression_done) {
-        ngx_chain_t  *compressed_chain;
-        ngx_chain_t  *decompressed_chain;
-        ngx_int_t     decompress_rc;
-        ngx_int_t     rc;
-
-        ngx_log_debug2(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                      "markdown filter: starting decompression, type=%d, size=%uz bytes",
-                      ctx->compression_type, ctx->buffer.size);
-
-        ctx->compressed_size = ctx->buffer.size;
-        rc = ngx_http_markdown_prepare_compressed_chain(r, ctx, conf, &compressed_chain);
-        if (rc != NGX_OK) {
-            return rc;
-        }
-
-        decompress_rc = ngx_http_markdown_decompress(r, ctx->compression_type,
-                                                      compressed_chain,
-                                                      &decompressed_chain);
-
-        if (decompress_rc == NGX_DECLINED) {
-            ngx_log_debug0(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                          "markdown filter: decompression not supported, "
-                          "returning original content (fail-open)");
-
-            return ngx_http_markdown_fail_open_buffered_response(r, ctx, NULL);
-        }
-
-        if (decompress_rc != NGX_OK) {
-            const ngx_str_t *compression_name;
-
-            compression_name = ngx_http_markdown_compression_name(
-                ctx->compression_type);
-
-            ngx_log_error(NGX_LOG_ERR, r->connection->log, 0,
-                         "markdown filter: decompression failed, compression=%V, "
-                         "error=\"decompression error\", category=conversion",
-                         compression_name);
-
-            ngx_atomic_fetch_add(&ngx_http_markdown_metrics.decompressions_failed, 1);
-            return ngx_http_markdown_reject_or_fail_open_buffered_response(
-                r, ctx, conf,
-                "markdown filter: fail-open strategy - returning original content");
-        }
-
-        rc = ngx_http_markdown_apply_decompressed_payload(r, ctx, conf, decompressed_chain);
-        if (rc != NGX_OK) {
-            return rc;
-        }
-
-        ngx_http_markdown_record_decompression_success(r, ctx);
-    }
-
-    return NGX_OK;
-}
-
-static ngx_int_t
-ngx_http_markdown_body_filter_convert_and_output(ngx_http_request_t *r,
-                                                 ngx_http_markdown_ctx_t *ctx,
-                                                 ngx_http_markdown_conf_t *conf)
-{
-    ngx_int_t             rc;
-    ngx_msec_t            elapsed_ms;
-    ngx_flag_t            has_result;
-    struct MarkdownResult result;
-
-    ctx->conversion_attempted = 1;
-    ngx_atomic_fetch_add(&ngx_http_markdown_metrics.conversions_attempted, 1);
-    elapsed_ms = 0;
-
-    ngx_log_debug1(NGX_LOG_DEBUG_HTTP, r->connection->log, 0,
-                  "markdown filter: buffered complete response, size: %uz bytes",
-                  ctx->buffer.size);
-
-    rc = ngx_http_markdown_resolve_conditional_result(
-        r, ctx, conf, &result, &elapsed_ms, &has_result);
-    if (rc == NGX_DONE) {
-        return NGX_OK;
-    }
-    if (rc != NGX_OK) {
-        return rc;
-    }
-
-    if (!has_result) {
-        rc = ngx_http_markdown_execute_conversion(r, ctx, conf, &result, &elapsed_ms);
-        if (rc != NGX_OK) {
-            return rc;
-        }
-    }
-
-    return ngx_http_markdown_send_conversion_output(r, ctx, conf, &result, elapsed_ms);
-}
-
-/*
- * Body filter
- *
- * Called for each chunk of the response body.
- * Buffers the response and performs conversion when complete.
- * 
- * This implements Task 14.7: Body filter hook
- * - Accumulates response chunks in buffer
- * - Detects when all chunks are buffered (last_buf flag)
- * - Calls Rust conversion engine via FFI
- * - Updates response headers on success
- * - Sends converted Markdown response
- * - Handles errors with configured strategy
- *
- * Requirements: FR-02.4, FR-04.1, FR-09.1, FR-09.2, FR-10.1, FR-10.3
- * 
- * @param r   The request structure
- * @param in  The input chain containing response body chunks
- * @return    NGX_OK on success, NGX_ERROR on error
- */
-static ngx_int_t
-ngx_http_markdown_body_filter(ngx_http_request_t *r, ngx_chain_t *in)
-{
-    ngx_http_markdown_ctx_t   *ctx;
-    ngx_http_markdown_conf_t  *conf;
-    ngx_int_t                  rc;
-
-    /* Get module configuration */
-    conf = ngx_http_get_module_loc_conf(r, ngx_http_markdown_filter_module);
-    if (conf == NULL) {
-        /* Module disabled, pass through */
-        return ngx_http_next_body_filter(r, in);
-    }
-
-    /*
-     * Get request context created by header filter.
-     *
-     * IMPORTANT: Do not re-evaluate markdown_filter here. Dynamic expressions
-     * can resolve differently between phases; body filter must follow the
-     * cached header-phase decision.
-     */
-    ctx = ngx_http_get_module_ctx(r, ngx_http_markdown_filter_module);
-    if (ctx == NULL) {
-        /* No context means header filter didn't set up conversion */
-        /* Pass through unchanged */
-        return ngx_http_next_body_filter(r, in);
-    }
-
-    if (!ctx->filter_enabled) {
-        r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-        return ngx_http_next_body_filter(r, in);
-    }
-
-    /* If not eligible for conversion, pass through */
-    if (!ctx->eligible) {
-        r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-        /* Track bypassed request (ineligible for conversion) */
-        ngx_atomic_fetch_add(&ngx_http_markdown_metrics.conversions_bypassed, 1);
-        if (ngx_http_markdown_forward_headers(r, ctx) != NGX_OK) {
-            return NGX_ERROR;
-        }
-        return ngx_http_next_body_filter(r, in);
-    }
-
-    /* If conversion already attempted, pass through */
-    if (ctx->conversion_attempted) {
-        r->buffered &= ~NGX_HTTP_MARKDOWN_BUFFERED;
-        return ngx_http_next_body_filter(r, in);
-    }
-
-    rc = ngx_http_markdown_body_filter_buffer_input(r, in, ctx, conf);
-    if (rc == NGX_AGAIN) {
-        return NGX_OK;
-    }
-    if (rc != NGX_OK) {
-        return rc;
-    }
-
-    rc = ngx_http_markdown_body_filter_decompress_if_needed(r, ctx, conf);
-    if (rc != NGX_OK) {
-        return rc;
-    }
-
-    return ngx_http_markdown_body_filter_convert_and_output(r, ctx, conf);
-}
-
-/*
- * Metrics endpoint handler
- *
- * This handler responds to requests for the metrics endpoint.
- * It returns metrics in plain text or JSON format based on the Accept header.
- * Access is restricted to localhost (127.0.0.1, ::1) by default for security.
- *
- * Response formats:
- * - Plain text (default): Accept: text/plain or no Accept header
- * - JSON: Accept: application/json
- *
- * Security:
- * - Only accessible from localhost by default
- * - Use NGINX allow/deny directives to customize access control
- *
- * Example configuration:
- *   location /markdown-metrics {
- *       markdown_metrics;
- *       allow 127.0.0.1;
- *       allow ::1;
- *       deny all;
- *   }
- *
- * @param r  The request structure
- * @return   NGX_OK on success, NGX_HTTP_* error code on failure
- *
- * Requirements: FR-13.5
- * Task: 22.4 Implement metrics exposure endpoint
- */
-static ngx_int_t
-ngx_http_markdown_metrics_handler(ngx_http_request_t *r)
-{
-    ngx_int_t     rc;
-    ngx_buf_t    *b;
-    ngx_chain_t   out;
-    u_char       *p;
-    size_t        len;
-    ngx_flag_t    json_format;
-    
-    /* Only allow GET and HEAD methods */
-    if (!(r->method & (NGX_HTTP_GET|NGX_HTTP_HEAD))) {
-        return NGX_HTTP_NOT_ALLOWED;
-    }
-    
-    /*
-     * Security: Restrict access to localhost by default
-     * 
-     * Check if the request is from localhost (127.0.0.1 or ::1).
-     * This provides a secure default, but administrators can override
-     * this using NGINX's allow/deny directives in the location block.
-     * 
-     * Note: This is a basic check. For production use, administrators
-     * should configure explicit allow/deny rules in the location block.
-     */
-    if (r->connection->sockaddr->sa_family == AF_INET) {
-        struct sockaddr_in *sin = (struct sockaddr_in *) r->connection->sockaddr;
-        /* Check for 127.0.0.1 (localhost) */
-        if (ntohl(sin->sin_addr.s_addr) != INADDR_LOOPBACK) {
-            ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                         "markdown_metrics: access denied from non-localhost IPv4 address");
-            return NGX_HTTP_FORBIDDEN;
-        }
-    }
-#if (NGX_HAVE_INET6)
-    else if (r->connection->sockaddr->sa_family == AF_INET6) {
-        struct sockaddr_in6 *sin6 = (struct sockaddr_in6 *) r->connection->sockaddr;
-        /* Check for ::1 (localhost) */
-        if (!IN6_IS_ADDR_LOOPBACK(&sin6->sin6_addr)) {
-            ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                         "markdown_metrics: access denied from non-localhost IPv6 address");
-            return NGX_HTTP_FORBIDDEN;
-        }
-    }
-#endif
-    else {
-        /* Unknown address family - deny by default */
-        ngx_log_error(NGX_LOG_WARN, r->connection->log, 0,
-                     "markdown_metrics: access denied from unknown address family");
-        return NGX_HTTP_FORBIDDEN;
-    }
-    
-    /*
-     * Determine output format based on Accept header
-     * 
-     * - Accept: application/json -> JSON format
-     * - Accept: text/plain or no Accept header -> Plain text format (default)
-     */
-    json_format = 0;
-    {
-        ngx_table_elt_t *accept_header = NULL;
-
-#if (NGX_HTTP_HEADERS)
-        accept_header = r->headers_in.accept;
-#else
-        static ngx_str_t  accept_name = ngx_string("Accept");
-
-        accept_header = ngx_http_markdown_find_request_header(
-            r, &accept_name);
-#endif
-
-        if (accept_header != NULL
-            && ngx_strstr(accept_header->value.data, "application/json") != NULL)
-        {
-            json_format = 1;
-        }
-    }
-    
-    /*
-     * Discard request body if present
-     * 
-     * The metrics endpoint doesn't need the request body, so we discard it.
-     * This is a standard pattern for GET endpoints in NGINX.
-     */
-    rc = ngx_http_discard_request_body(r);
-    if (rc != NGX_OK) {
-        return rc;
-    }
-    
-    /*
-     * Allocate buffer for response
-     * 
-     * We allocate a buffer large enough to hold the metrics output.
-     * The buffer size is estimated based on the format:
-     * - Plain text: ~500 bytes
-     * - JSON: ~800 bytes
-     * 
-     * We allocate 2048 bytes to be safe and allow for future expansion.
-     */
-    b = ngx_create_temp_buf(r->pool, 2048);
-    if (b == NULL) {
-        ngx_log_error(NGX_LOG_CRIT, r->connection->log, 0,
-                     "markdown_metrics: failed to allocate response buffer");
-        return NGX_HTTP_INTERNAL_SERVER_ERROR;
-    }
-    
-    /*
-     * Format metrics output
-     * 
-     * Read metrics from the global metrics structure and format them
-     * according to the requested format (plain text or JSON).
-     */
-    p = b->pos;
-    
-    if (json_format) {
-        /*
-         * JSON format
-         * 
-         * Format metrics as JSON object with all fields.
-         * This format is machine-readable and suitable for monitoring tools.
-         * 
-         * Example output:
-         * {
-         *   "conversions_attempted": 100,
-         *   "conversions_succeeded": 95,
-         *   "conversions_failed": 3,
-         *   "conversions_bypassed": 2,
-         *   "failures_conversion": 2,
-         *   "failures_resource_limit": 1,
-         *   "failures_system": 0,
-         *   "conversion_time_sum_ms": 1250,
-         *   "input_bytes": 524288,
-         *   "output_bytes": 131072,
-         *   "decompressions_attempted": 50,
-         *   "decompressions_succeeded": 48,
-         *   "decompressions_failed": 2,
-         *   "decompressions_gzip": 40,
-         *   "decompressions_deflate": 8,
-         *   "decompressions_brotli": 0
-         * }
-         */
-        p = ngx_slprintf(p, b->end,
-            "{\n"
-            "  \"conversions_attempted\": %uA,\n"
-            "  \"conversions_succeeded\": %uA,\n"
-            "  \"conversions_failed\": %uA,\n"
-            "  \"conversions_bypassed\": %uA,\n"
-            "  \"failures_conversion\": %uA,\n"
-            "  \"failures_resource_limit\": %uA,\n"
-            "  \"failures_system\": %uA,\n"
-            "  \"conversion_time_sum_ms\": %uA,\n"
-            "  \"input_bytes\": %uA,\n"
-            "  \"output_bytes\": %uA,\n"
-            "  \"decompressions_attempted\": %uA,\n"
-            "  \"decompressions_succeeded\": %uA,\n"
-            "  \"decompressions_failed\": %uA,\n"
-            "  \"decompressions_gzip\": %uA,\n"
-            "  \"decompressions_deflate\": %uA,\n"
-            "  \"decompressions_brotli\": %uA\n"
-            "}",
-            ngx_http_markdown_metrics.conversions_attempted,
-            ngx_http_markdown_metrics.conversions_succeeded,
-            ngx_http_markdown_metrics.conversions_failed,
-            ngx_http_markdown_metrics.conversions_bypassed,
-            ngx_http_markdown_metrics.failures_conversion,
-            ngx_http_markdown_metrics.failures_resource_limit,
-            ngx_http_markdown_metrics.failures_system,
-            ngx_http_markdown_metrics.conversion_time_sum_ms,
-            ngx_http_markdown_metrics.input_bytes,
-            ngx_http_markdown_metrics.output_bytes,
-            ngx_http_markdown_metrics.decompressions_attempted,
-            ngx_http_markdown_metrics.decompressions_succeeded,
-            ngx_http_markdown_metrics.decompressions_failed,
-            ngx_http_markdown_metrics.decompressions_gzip,
-            ngx_http_markdown_metrics.decompressions_deflate,
-            ngx_http_markdown_metrics.decompressions_brotli);
-    } else {
-        /*
-         * Plain text format
-         * 
-         * Format metrics as human-readable plain text with labels.
-         * This format is easy to read and suitable for manual inspection.
-         * 
-         * Example output:
-         * Markdown Filter Metrics
-         * =======================
-         * Conversions Attempted: 100
-         * Conversions Succeeded: 95
-         * Conversions Failed: 3
-         * Conversions Bypassed: 2
-         * 
-         * Failure Breakdown:
-         * - Conversion Errors: 2
-         * - Resource Limit Exceeded: 1
-         * - System Errors: 0
-         * 
-         * Performance:
-         * - Total Conversion Time: 1250 ms
-         * - Total Input Bytes: 524288
-         * - Total Output Bytes: 131072
-         * 
-         * Decompression Statistics:
-         * - Decompressions Attempted: 50
-         * - Decompressions Succeeded: 48
-         * - Decompressions Failed: 2
-         * - Gzip Decompressions: 40
-         * - Deflate Decompressions: 8
-         * - Brotli Decompressions: 0
-         */
-        p = ngx_slprintf(p, b->end,
-            "Markdown Filter Metrics\n"
-            "=======================\n"
-            "Conversions Attempted: %uA\n"
-            "Conversions Succeeded: %uA\n"
-            "Conversions Failed: %uA\n"
-            "Conversions Bypassed: %uA\n"
-            "\n"
-            "Failure Breakdown:\n"
-            "- Conversion Errors: %uA\n"
-            "- Resource Limit Exceeded: %uA\n"
-            "- System Errors: %uA\n"
-            "\n"
-            "Performance:\n"
-            "- Total Conversion Time: %uA ms\n"
-            "- Total Input Bytes: %uA\n"
-            "- Total Output Bytes: %uA\n"
-            "\n"
-            "Decompression Statistics:\n"
-            "- Decompressions Attempted: %uA\n"
-            "- Decompressions Succeeded: %uA\n"
-            "- Decompressions Failed: %uA\n"
-            "- Gzip Decompressions: %uA\n"
-            "- Deflate Decompressions: %uA\n"
-            "- Brotli Decompressions: %uA\n",
-            ngx_http_markdown_metrics.conversions_attempted,
-            ngx_http_markdown_metrics.conversions_succeeded,
-            ngx_http_markdown_metrics.conversions_failed,
-            ngx_http_markdown_metrics.conversions_bypassed,
-            ngx_http_markdown_metrics.failures_conversion,
-            ngx_http_markdown_metrics.failures_resource_limit,
-            ngx_http_markdown_metrics.failures_system,
-            ngx_http_markdown_metrics.conversion_time_sum_ms,
-            ngx_http_markdown_metrics.input_bytes,
-            ngx_http_markdown_metrics.output_bytes,
-            ngx_http_markdown_metrics.decompressions_attempted,
-            ngx_http_markdown_metrics.decompressions_succeeded,
-            ngx_http_markdown_metrics.decompressions_failed,
-            ngx_http_markdown_metrics.decompressions_gzip,
-            ngx_http_markdown_metrics.decompressions_deflate,
-            ngx_http_markdown_metrics.decompressions_brotli);
-    }
-    
-    /* Calculate actual content length */
-    len = p - b->pos;
-    b->last = p;
-    
-    /*
-     * Set response headers
-     * 
-     * - Status: 200 OK
-     * - Content-Type: application/json or text/plain
-     * - Content-Length: actual length of the response body
-     */
-    r->headers_out.status = NGX_HTTP_OK;
-    r->headers_out.content_length_n = len;
-    
-    if (json_format) {
-        ngx_str_set(&r->headers_out.content_type, "application/json");
-    } else {
-        ngx_str_set(&r->headers_out.content_type, "text/plain");
-    }
-    r->headers_out.content_type_len = r->headers_out.content_type.len;
-    
-    /*
-     * Mark buffer as last in chain
-     * 
-     * This tells NGINX that this is the final buffer in the response.
-     */
-    b->last_buf = (r == r->main) ? 1 : 0;
-    b->last_in_chain = 1;
-    
-    /*
-     * Send response headers
-     */
-    rc = ngx_http_send_header(r);
-    if (rc == NGX_ERROR || rc > NGX_OK || r->header_only) {
-        return rc;
-    }
-    
-    /*
-     * Send response body
-     * 
-     * For HEAD requests, we don't send the body (r->header_only is set).
-     * For GET requests, we send the formatted metrics.
-     */
-    out.buf = b;
-    out.next = NULL;
-    
-    return ngx_http_output_filter(r, &out);
-}
+#include "ngx_http_markdown_metrics_impl.h"
 
 /*
  * Find a request header by name in the generic headers list.
@@ -3188,106 +148,3 @@ ngx_http_markdown_find_request_header(ngx_http_request_t *r, const ngx_str_t *na
     return NULL;
 }
 #endif
-
-static ngx_int_t
-ngx_http_markdown_forward_headers(ngx_http_request_t *r, ngx_http_markdown_ctx_t *ctx)
-{
-    ngx_int_t rc;
-
-    if (ctx == NULL || ctx->headers_forwarded) {
-        return NGX_OK;
-    }
-
-    rc = ngx_http_next_header_filter(r);
-    if (rc == NGX_ERROR || rc > NGX_OK) {
-        return rc;
-    }
-
-    ctx->headers_forwarded = 1;
-    return rc;
-}
-
-/*
- * Send the fully buffered original HTML response (fail-open path).
- *
- * This is used when conversion fails after the module has already buffered and
- * consumed the entire upstream body. At that point, forwarding the current
- * input chain would lose data; we must emit the buffered original body.
- */
-static ngx_int_t
-ngx_http_markdown_send_buffered_original_response(ngx_http_request_t *r,
-    ngx_http_markdown_ctx_t *ctx)
-{
-    ngx_buf_t    *b;
-    ngx_chain_t  *out;
-
-    if (ctx == NULL) {
-        return NGX_ERROR;
-    }
-
-    b = ngx_pcalloc(r->pool, sizeof(ngx_buf_t));
-    if (b == NULL) {
-        return NGX_ERROR;
-    }
-
-    if (r->method == NGX_HTTP_HEAD || ctx->buffer.size == 0) {
-        b->pos = NULL;
-        b->last = NULL;
-        b->memory = 0;
-    } else {
-        b->pos = ctx->buffer.data;
-        b->last = ctx->buffer.data + ctx->buffer.size;
-        b->memory = 1;
-    }
-
-    b->last_buf = (r == r->main) ? 1 : 0;
-    b->last_in_chain = 1;
-
-    out = ngx_alloc_chain_link(r->pool);
-    if (out == NULL) {
-        return NGX_ERROR;
-    }
-    out->buf = b;
-    out->next = NULL;
-
-    return ngx_http_next_body_filter(r, out);
-}
-
-/*
- * Fail-open while buffering is in progress by replaying the already-buffered
- * prefix and then forwarding the unconsumed upstream chain.
- *
- * This preserves correctness when a size limit is exceeded after some chunks
- * were already copied into the module buffer and marked consumed.
- */
-static ngx_int_t
-ngx_http_markdown_fail_open_with_buffered_prefix(ngx_http_request_t *r,
-    ngx_http_markdown_ctx_t *ctx, ngx_chain_t *remaining)
-{
-    ngx_buf_t    *b;
-    ngx_chain_t  *out;
-
-    if (ctx == NULL || ctx->buffer.size == 0) {
-        return ngx_http_next_body_filter(r, remaining);
-    }
-
-    b = ngx_pcalloc(r->pool, sizeof(ngx_buf_t));
-    if (b == NULL) {
-        return NGX_ERROR;
-    }
-
-    b->pos = ctx->buffer.data;
-    b->last = ctx->buffer.data + ctx->buffer.size;
-    b->memory = 1;
-    b->last_buf = 0;
-    b->last_in_chain = 0;
-
-    out = ngx_alloc_chain_link(r->pool);
-    if (out == NULL) {
-        return NGX_ERROR;
-    }
-    out->buf = b;
-    out->next = remaining;
-
-    return ngx_http_next_body_filter(r, out);
-}
