+++ components/nginx-module/tests/unit/metrics_endpoint_test.c
@@ -9,6 +9,11 @@ typedef struct {
     unsigned long conversions_attempted;
     unsigned long conversions_succeeded;
     unsigned long conversions_failed;
+    unsigned long conversion_time_sum_ms;
+    unsigned long conversion_latency_le_10ms;
+    unsigned long conversion_latency_le_100ms;
+    unsigned long conversion_latency_le_1000ms;
+    unsigned long conversion_latency_gt_1000ms;
     unsigned long decompressions_attempted;
     unsigned long decompressions_succeeded;
 } metrics_t;
@@ -28,6 +33,8 @@ static endpoint_response_t
 handle_metrics_request(const char *method, const char *remote_addr, const char *format, const metrics_t *m)
 {
     endpoint_response_t out;
+    unsigned long completed;
+    unsigned long avg_ms;
     memset(&out, 0, sizeof(out));
 
     if (!(STR_EQ(method, "GET") || STR_EQ(method, "HEAD"))) {
@@ -40,20 +47,38 @@ handle_metrics_request(const char *method, const char *remote_addr, const char *
     }
 
     out.status = 200;
+    completed = m->conversions_succeeded + m->conversions_failed;
+    avg_ms = completed == 0 ? 0 : m->conversion_time_sum_ms / completed;
     if (STR_EQ(format, "json")) {
         snprintf(out.body, sizeof(out.body),
                  "{\"conversions_attempted\":%lu,\"conversions_succeeded\":%lu,\"conversions_failed\":%lu,"
+                 "\"conversion_completed\":%lu,\"conversion_time_avg_ms\":%lu,"
+                 "\"conversion_latency_buckets\":{\"le_10ms\":%lu,\"le_100ms\":%lu,\"le_1000ms\":%lu,\"gt_1000ms\":%lu},"
                  "\"decompressions_attempted\":%lu,\"decompressions_succeeded\":%lu}",
                  m->conversions_attempted, m->conversions_succeeded, m->conversions_failed,
+                 completed, avg_ms,
+                 m->conversion_latency_le_10ms, m->conversion_latency_le_100ms,
+                 m->conversion_latency_le_1000ms, m->conversion_latency_gt_1000ms,
                  m->decompressions_attempted, m->decompressions_succeeded);
     } else {
         snprintf(out.body, sizeof(out.body),
-                 "markdown_conversions_attempted_total %lu\n"
-                 "markdown_conversions_succeeded_total %lu\n"
-                 "markdown_conversions_failed_total %lu\n"
-                 "markdown_decompressions_attempted_total %lu\n"
-                 "markdown_decompressions_succeeded_total %lu\n",
+                 "Markdown Filter Metrics\n"
+                 "=======================\n"
+                 "Conversions Attempted: %lu\n"
+                 "Conversions Succeeded: %lu\n"
+                 "Conversions Failed: %lu\n"
+                 "Conversions Completed: %lu\n"
+                 "Average Conversion Time: %lu ms\n"
+                 "Latency <= 10ms: %lu\n"
+                 "Latency <= 100ms: %lu\n"
+                 "Latency <= 1000ms: %lu\n"
+                 "Latency > 1000ms: %lu\n"
+                 "Decompressions Attempted: %lu\n"
+                 "Decompressions Succeeded: %lu\n",
                  m->conversions_attempted, m->conversions_succeeded, m->conversions_failed,
+                 completed, avg_ms,
+                 m->conversion_latency_le_10ms, m->conversion_latency_le_100ms,
+                 m->conversion_latency_le_1000ms, m->conversion_latency_gt_1000ms,
                  m->decompressions_attempted, m->decompressions_succeeded);
     }
     return out;
@@ -66,6 +91,11 @@ sample_metrics(void)
     m.conversions_attempted = 10;
     m.conversions_succeeded = 8;
     m.conversions_failed = 2;
+    m.conversion_time_sum_ms = 100;
+    m.conversion_latency_le_10ms = 4;
+    m.conversion_latency_le_100ms = 5;
+    m.conversion_latency_le_1000ms = 1;
+    m.conversion_latency_gt_1000ms = 0;
     m.decompressions_attempted = 4;
     m.decompressions_succeeded = 3;
     return m;
@@ -105,12 +135,16 @@ test_output_formats(void)
     TEST_SUBSECTION("Plain text and JSON output formats");
 
     r = handle_metrics_request("GET", "::1", "text", &m);
-    TEST_ASSERT(strstr(r.body, "markdown_conversions_attempted_total 10") != NULL,
+    TEST_ASSERT(strstr(r.body, "Conversions Attempted: 10") != NULL,
                 "Plain text output should include counters");
+    TEST_ASSERT(strstr(r.body, "Average Conversion Time: 10 ms") != NULL,
+                "Plain text output should include averages");
 
     r = handle_metrics_request("GET", "::1", "json", &m);
     TEST_ASSERT(strstr(r.body, "\"conversions_attempted\":10") != NULL,
                 "JSON output should include counters");
+    TEST_ASSERT(strstr(r.body, "\"conversion_latency_buckets\":{") != NULL,
+                "JSON output should include latency buckets");
     TEST_ASSERT(strstr(r.body, "\"decompressions_succeeded\":3") != NULL,
                 "JSON output should include decompression counters");
     TEST_PASS("Output formats work");
